Fixing and testing Root injection
