+++ boson-core/src/main/scala/io/zink/boson/bson/bsonImpl/BosonImpl.scala
@@ -5,6 +5,7 @@ import java.time.Instant
 
 import io.netty.buffer.{ByteBuf, Unpooled}
 import io.zink.boson.bson.bsonImpl.Dictionary._
+import io.zink.boson.bson.bsonPath._
 import io.zink.boson.bson.codec._
 
 import scala.util.{Failure, Success, Try}
@@ -39,6 +40,10 @@ class BosonImpl(
                  javaByteBuf: Option[ByteBuffer] = None,
                  stringJson: Option[String] = None
                ) {
+
+
+  type DataStructure = Either[ByteBuf, String]
+  type StatementsList = List[(Statement, String)]
 
   /**
     * Deprecated, used on previous implementations of Boson to verify which input was given by the user.
@@ -1060,6 +1065,664 @@ class BosonImpl(
   @deprecated
   def getByteBuf: Either[ByteBuf, String] = this.nettyBuffer
 
+
+  //----------------------------------------------------- INJECTORS ----------------------------------------------------
+
+
+  /**
+    * Starter method for the injection process, this method will pattern match the statements in the statements list
+    * and will delegate to other helper methods
+    *
+    * @param dataStructure - The data structure in which to perform the injection process (either a ByteBuf or a String)
+    * @param statements    - The statements with information regarding where to perform the injection
+    * @param injFunction   - The injection function to be applied
+    * @tparam T - The type of the input and output of the injection function
+    * @return a new codec with the changes applied to it
+    */
+  def inject[T](dataStructure: DataStructure, statements: StatementsList, injFunction: T => T): Codec = {
+    val codec: Codec = dataStructure match {
+      case Right(jsonString) => CodecObject.toCodec(jsonString)
+      case Left(byteBuf) => CodecObject.toCodec(byteBuf)
+    }
+
+    statements.head._1 match {
+      case ROOT => rootInjection(codec, injFunction)
+
+      case Key(key: String) => modifyAll(statements, codec, key, injFunction) // (key = fieldID)
+
+      case HalfName(half: String) => modifyAll(statements, codec, half, injFunction)
+
+      case HasElem(key: String, elem: String) => modifyHasElem(statements, codec, key, elem, injFunction)
+
+      case ArrExpr(leftArg: Int, midArg: Option[RangeCondition], rightArg: Option[Any]) => ???
+
+      case KeyWithArrExpr(key: String, arrEx: ArrExpr) =>
+        val input: (String, Int, String, Any) =
+          (arrEx.leftArg, arrEx.midArg, arrEx.rightArg) match {
+
+            case (_, o1, o2) if o1.isDefined && o2.isDefined =>
+              (key, arrEx.leftArg, o1.get.value, o2.get) //User sent, for example, Key[1 TO 3] translate to - Key[1 TO 3]
+
+            case (_, o1, o2) if o1.isEmpty && o2.isEmpty =>
+              (key, arrEx.leftArg, TO_RANGE, arrEx.leftArg) //User sent, for example, Key[1] translates to - Key[1 TO 1]
+
+            case (0, str, None) =>
+              str.get.value match {
+                case C_FIRST =>
+                  (key, 0, TO_RANGE, 0) //User sent Key[first] , translates to - Key[0 TO 0]
+                case C_END =>
+                  (key, 0, C_END, None)
+                case C_ALL =>
+                  (key, 0, TO_RANGE, C_END) //user sent Key[all], translates to - Key[0 TO END]
+              }
+          }
+        val modifiedCodec = arrayInjection(statements, codec, codec.duplicate, injFunction, input._1, input._2, input._3, input._4)
+        codec + modifiedCodec
+
+      case _ => throw CustomException("Wrong Statements, Bad Expression.")
+    }
+  }
+
+  /**
+    *
+    * @param statementsList -
+    * @param codec          -
+    * @param fieldID        -
+    * @param injFunction    -
+    * @tparam T -
+    * @return
+    */
+  private def modifyAll[T](statementsList: StatementsList, codec: Codec, fieldID: String, injFunction: T => T): Codec = {
+    def writeCodec(currentCodec: Codec, startReader: Int, originalSize: Int): Codec = {
+      if ((codec.getReaderIndex - startReader) < originalSize) currentCodec
+      else {
+        val dataType: Int = codec.readDataType
+        val codecWithDataType = codec.writeToken(currentCodec, SonNumber(CS_BYTE, dataType))
+        val newCodec = dataType match {
+          case 0 => writeCodec(codecWithDataType, startReader, originalSize) // This is the end
+          case _ =>
+            val (codecWithKey, key) = writeKeyAndByte(codec, codecWithDataType)
+
+            key match {
+              case extracted if fieldID.toCharArray.deep == extracted.toCharArray.deep => //add isHalWord Later
+                if (statementsList.lengthCompare(1) == 0) {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT)) {
+                    dataType match {
+                      case D_BSONOBJECT | D_BSONARRAY =>
+                        val token = if (dataType == D_BSONOBJECT) SonObject(CS_OBJECT) else SonArray(CS_ARRAY)
+                        val partialData = codec.readToken(token) match {
+                          case SonObject(_, result) => result match {
+                            case byteBuf: ByteBuf => Left(byteBuf)
+                            case string: String => Right(string)
+                          }
+                          case SonArray(_, result) => result match {
+                            case byteBuf: ByteBuf => Left(byteBuf)
+                            case string: String => Right(string)
+                          }
+                        }
+                        val partialCodec = inject(partialData, statementsList, injFunction)
+                        modifierAll(codec, codecWithKey + partialCodec, dataType, injFunction)
+                      case _ =>
+                        modifierAll(codec, codecWithKey, dataType, injFunction)
+                    }
+                  } else {
+                    modifierAll(codec, codecWithKey, dataType, injFunction)
+                  }
+                } else {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT)) {
+                    dataType match {
+                      case D_BSONOBJECT | D_BSONARRAY =>
+                        val token = if (dataType == D_BSONOBJECT) SonObject(CS_OBJECT) else SonArray(CS_ARRAY)
+                        val partialData = codec.readToken(token) match {
+                          case SonObject(_, result) => result match {
+                            case byteBuf: ByteBuf => Left(byteBuf)
+                            case string: String => Right(string)
+                          }
+                          case SonArray(_, result) => result match {
+                            case byteBuf: ByteBuf => Left(byteBuf)
+                            case string: String => Right(string)
+                          }
+                        }
+                        val modifiedPartialCodec = inject(partialData, statementsList.drop(1), injFunction)
+                        inject((codecWithKey + modifiedPartialCodec).getCodecData, statementsList, injFunction)
+
+                      case _ =>
+                        processTypesAll(statementsList, dataType, codec, codecWithKey, fieldID, injFunction)
+                    }
+                  } else {
+                    dataType match {
+                      case D_BSONOBJECT | D_BSONARRAY =>
+                        inject(codec.getCodecData, statementsList.drop(1), injFunction)
+                      case _ =>
+                        processTypesArray(dataType, codec, codecWithKey)
+                    }
+                  }
+                }
+              case x if fieldID.toCharArray.deep != x.toCharArray.deep => //TODO - add !isHalfWord
+                if (statementsList.head._2.contains(C_DOUBLEDOT))
+                  processTypesAll(statementsList, dataType, codec, codecWithKey, fieldID, injFunction)
+                else
+                  processTypesArray(dataType, codec, codecWithKey)
+            }
+        }
+        writeCodec(newCodec, startReader, originalSize)
+      }
+    }
+
+    val startReader: Int = codec.getReaderIndex
+    val originalSize: Int = codec.readSize
+    //TODO DONT FORGET TO WRITE THE SIZE TO THE RESULT CODEC
+    val emptyCodec: Codec = codec.getCodecData match {
+      case Left(_) => CodecObject.toCodec(Unpooled.EMPTY_BUFFER)
+      case Right(_) => CodecObject.toCodec("")
+    }
+
+    val codecWithoutSize = writeCodec(emptyCodec, startReader, originalSize)
+    val finalSize = codecWithoutSize.getCodecData match {
+      case Left(byteBuf) => byteBuf.capacity
+      case Right(string) => string.length
+    }
+
+    emptyCodec.writeToken(emptyCodec, SonNumber(CS_INTEGER, finalSize)) + codecWithoutSize
+  }
+
+  /**
+    * Function used to search for a element within an object
+    *
+    * @param statementsList - A list with pairs that contains the key of interest and the type of operation
+    * @param codec          - Structure from which we are reading the old values
+    * @param fieldID        - Name of the field of interest
+    * @param elem           - Name of the element to look for inside the objects inside an Array
+    * @param injFunction    - The injection function to be applied
+    * @tparam T - The type of input and output of the injection function
+    * @return a modified Codec where the injection function may have been applied to the desired element (if it exists)
+    */
+  private def modifyHasElem[T](statementsList: StatementsList, codec: Codec, fieldID: String, elem: String, injFunction: T => T): Codec = {
+
+    val startReader: Int = codec.getReaderIndex
+    val originalSize: Int = codec.readSize
+    //TODO DONT FORGET TO WRITE THE SIZE TO THE RESULT CODEC
+    val emptyDataStructure: Codec = codec.getCodecData match {
+      case Left(_) => CodecObject.toCodec(Unpooled.EMPTY_BUFFER)
+      case Right(_) => CodecObject.toCodec("")
+    }
+
+    /**
+      * Recursive function to iterate through the given data structure and return the modified codec
+      *
+      * @param writableCodec - the codec to be modified
+      * @return A modified codec
+      */
+    def iterateDataStructure(writableCodec: Codec): Codec = {
+      if ((codec.getReaderIndex - startReader) < originalSize) writableCodec
+      else {
+        val dataType: Int = codec.readDataType
+        val codecWithDataType = codec.writeToken(writableCodec, SonNumber(CS_BYTE, dataType)) //write the read byte to a Codec
+        dataType match {
+          case 0 => iterateDataStructure(codecWithDataType)
+          case _ => //In case its not the end
+
+            val (codecWithKeyByte, key) = writeKeyAndByte(codec, codecWithDataType)
+
+            //We only want to modify if the dataType is an Array and if the extractedKey matches with the fieldID
+            //or they're halfword's
+            //in all other cases we just want to copy the data from one codec to the other (using "process" like functions)
+            key match {
+              case extracted if (fieldID.toCharArray.deep == extracted.toCharArray.deep || isHalfword(fieldID, extracted)) && dataType == D_BSONARRAY =>
+                //the key is a halfword and matches with the extracted key, dataType is an array
+                val modifiedCodec: Codec = searchAndModify(statementsList, codec, fieldID, injFunction, codecWithKeyByte)
+                iterateDataStructure(modifiedCodec)
+
+              case _ =>
+                if (statementsList.head._2.contains(C_DOUBLEDOT)) {
+                  val processedCodec: Codec = processTypesHasElem(statementsList, dataType, fieldID, elem, codec, codecWithKeyByte, injFunction)
+                  iterateDataStructure(processedCodec)
+                } else {
+                  val processedCodec: Codec = processTypesArray(dataType, codec, codecWithKeyByte)
+                  iterateDataStructure(processedCodec)
+                }
+            }
+        }
+      }
+    }
+
+    iterateDataStructure(emptyDataStructure) //Initiate recursion with an empty data structure
+  }
+
+  /**
+    * Function used to search for an element inside an object inside an array after finding the key of interest
+    *
+    * @param statementsList - A list with pairs that contains the key of interest and the type of operation
+    * @param codec          - Structure from which we are reading the old values
+    * @param fieldID        - Name of the field of interest
+    * @param injFunction    - The injection function to be applied
+    * @param writableCodec  - Structure to where we write the values
+    * @tparam T - The type of input and output of the injection function
+    * @return a new Codec with the value injected
+    */
+  private def searchAndModify[T](statementsList: StatementsList, codec: Codec, fieldID: String, injFunction: T => T, writableCodec: Codec): Codec = {
+    val startReader: Int = codec.getReaderIndex
+    val originalSize: Int = codec.readSize
+
+    def iterateDataStructure(writtableCodec: Codec): Codec = {
+      if ((codec.getReaderIndex - startReader) < originalSize) writableCodec
+      else {
+        val dataType = codec.readDataType
+        val codecWithDataType = codec.writeToken(writableCodec, SonNumber(CS_BYTE, dataType)) //write the read byte to a Codec
+        dataType match {
+          case 0 => iterateDataStructure(codecWithDataType)
+          case _ =>
+            val (codecWithKey, key) = writeKeyAndByte(codec, codecWithDataType)
+            dataType match {
+              case D_BSONOBJECT =>
+                val bsonSize = codec.getSize
+                val partialCodec: Codec = codec.getToken(SonObject(CS_OBJECT_WITH_SIZE)) match {
+                  case SonObject(_, dataStruct) => dataStruct match {
+                    case byteBuf: ByteBuf => CodecObject.toCodec(byteBuf)
+                    case jsonString: String => CodecObject.toCodec(jsonString)
+                  }
+                } //Obtain only the size and the object itself of this BsonObject
+                if (hasElem(partialCodec.duplicate, fieldID)) {
+
+                  if (statementsList.size == 1) {
+
+                    val codecWithObj: Codec = codec.readSpecificSize(bsonSize)
+                    val byteArr: Array[Byte] = codecWithObj.getCodecData match {
+                      case Left(byteBuff) => byteBuff.array
+                      case Right(jsonString) => jsonString.getBytes
+                    }
+                    val newArray: Array[Byte] = applyFunction(injFunction, byteArr).asInstanceOf[Array[Byte]]
+                    val codecFromByteArr = codec.fromByteArray(newArray)
+                    if (statementsList.head._2.contains(C_DOUBLEDOT)) {
+                      val newCodec = inject(codec.getCodecData, statementsList, injFunction)
+                      iterateDataStructure(codecWithKey + newCodec)
+                    } else {
+                      iterateDataStructure(codecWithKey + codecFromByteArr)
+                    }
+
+                  } else {
+
+                    if (statementsList.head._2.contains(C_DOUBLEDOT)) {
+
+                      val dataStruct: Either[ByteBuf, String] = codec.getToken(SonString(CS_ARRAY)) match {
+                        case SonString(_, data) => data match {
+                          case byteBuf: ByteBuf => Left(byteBuf)
+                          case jsonString: String => Right(jsonString)
+                        }
+                      }
+                      val newCodec = inject(dataStruct, statementsList.drop(1), injFunction)
+                      val modifiedCodec = inject(newCodec.duplicate.getCodecData, statementsList, injFunction)
+                      iterateDataStructure(codecWithKey + modifiedCodec)
+
+                    } else {
+
+                      codec.readSpecificSize(bsonSize) //read the size of the object
+                      val newCodec = inject(partialCodec.getCodecData, statementsList.drop(1), injFunction)
+                      iterateDataStructure(codecWithKey + newCodec)
+
+                    }
+                  }
+
+                } else {
+                  val newCodec = codec.readSpecificSize(bsonSize) //read the size of the object
+                  iterateDataStructure(codecWithKey + newCodec)
+                }
+
+              case _ => processTypesArray(dataType, codec, writableCodec)
+            }
+        }
+      }
+      //TODO maybe do something more here ?
+    }
+
+    iterateDataStructure(writableCodec)
+  }
+
+  /**
+    * Method used to see if an object contains a certain element inside it
+    *
+    * @param codec - The structure in which to look for the element
+    * @param elem  - The name of the element to look for
+    * @return A boolean value saying if the given element is present in that object
+    */
+  private def hasElem(codec: Codec, elem: String): Boolean = {
+    val size: Int = codec.readSize
+    var key: String = ""
+    //Iterate through all of the keys from the dataStructure in order to see if it contains the elem
+    while (codec.getReaderIndex < size && (!elem.equals(key) && !isHalfword(elem, key))) {
+      key = "" //clear the key
+      val dataType = codec.readDataType
+      dataType match {
+        case 0 => //TODO what do we do in this case
+        case _ =>
+          val key: String = codec.readToken(SonString(CS_NAME)) match {
+            case SonString(_, keyString) => keyString.asInstanceOf[String]
+          }
+          codec.readToken(SonBoolean(C_ZERO)) //read the closing byte of the key, we're not interested in this value
+          dataType match {
+            case D_ZERO_BYTE => //TODO what do we do in this case
+
+            case D_FLOAT_DOUBLE => codec.readToken(SonNumber(CS_DOUBLE))
+
+            case D_ARRAYB_INST_STR_ENUM_CHRSEQ => codec.readToken(SonString(CS_ARRAY))
+
+            case D_BSONOBJECT | D_BSONARRAY => codec.getToken(SonString(CS_ARRAY))
+
+            case D_BOOLEAN => codec.readToken(SonBoolean(CS_BOOLEAN))
+
+            case D_NULL => //TODO what do we do in this case
+
+            case D_INT => codec.readToken(SonNumber(CS_INTEGER))
+
+            case D_LONG => codec.readToken(SonNumber(CS_LONG))
+          }
+      }
+    }
+    key.toCharArray.deep == elem.toCharArray.deep || isHalfword(elem, key)
+  }
+
+  /**
+    * Method that will perform the injection in the root of the data structure
+    *
+    * @param codec       - Codec encapsulating the data structure to inject in
+    * @param injFunction - The injection function to be applied
+    * @tparam T - The type of elements the injection function receives
+    * @return - A new codec with the injFunction applied to it
+    */
+  private def rootInjection[T](codec: Codec, injFunction: T => T): Codec = {
+    codec.getCodecData match {
+      case Left(byteBuf) =>
+        val bsonBytes: Array[Byte] = byteBuf.array() //extract the bytes from the bytebuf
+        bsonBytes.foreach(b => print(b + ", "))
+        val modifiedBytes: Array[Byte] = applyFunction(injFunction, bsonBytes).asInstanceOf[Array[Byte]] //apply the injector function to the extracted bytes
+      //        modifiedBytes.foreach(b => print(b + ", "))
+      val newBuf = Unpooled.buffer(modifiedBytes.length).writeBytes(modifiedBytes) //create a new ByteBuf from those bytes
+        CodecObject.toCodec(newBuf)
+
+      case Right(jsonString) => //TODO not sure if this is correct for CodecJson
+        val modifiedString: String = applyFunction(injFunction, jsonString).asInstanceOf[String]
+        CodecObject.toCodec(modifiedString)
+    }
+  }
+
+  /**
+    * Function used to copy values that aren't of interest while searching for a element inside a object inside a array
+    *
+    * @param statementsList - A list with pairs that contains the key of interest and the type of operation
+    * @param dataType       - Type of the value found
+    * @param fieldID        - Name of the field of interest
+    * @param elem           - Name of the element to search inside the objects inside an Array
+    * @param codec          - Structure from which we are reading the old values
+    * @param resultCodec    - Structure to where we write the values
+    * @param injFunction    - Injection function to be applied
+    * @tparam T - The type of input and output of the injection function
+    * @return a new Codec with the copied information
+    */
+  private def processTypesHasElem[T](statementsList: StatementsList, dataType: Int, fieldID: String, elem: String, codec: Codec, resultCodec: Codec, injFunction: T => T): Codec = dataType match {
+    case D_BSONOBJECT =>
+      val bsonObjectCodec: Codec = codec.getToken(SonString(CS_ARRAY)) match {
+        case SonString(_, data) => data match {
+          case byteBuff: ByteBuf => CodecObject.toCodec(byteBuff)
+          case jsonString: String => CodecObject.toCodec(jsonString)
+        }
+      }
+      val modifiedCodec: Codec = modifyHasElem(statementsList, bsonObjectCodec, fieldID, elem, injFunction)
+      resultCodec + modifiedCodec
+
+    case D_BSONARRAY =>
+      val length = codec.getSize
+      val partialCodec: Codec = codec.readSlice(length)
+      val modifiedCodec: Codec = modifyHasElem(statementsList, partialCodec, fieldID, elem, injFunction)
+      resultCodec + modifiedCodec
+
+    case D_FLOAT_DOUBLE => codec.writeToken(resultCodec, codec.readToken(SonNumber(CS_DOUBLE)))
+
+    case D_ARRAYB_INST_STR_ENUM_CHRSEQ => codec.writeToken(resultCodec, SonString(CS_ARRAY_WITH_SIZE))
+
+    case D_INT => codec.writeToken(resultCodec, SonNumber(CS_INTEGER))
+
+    case D_LONG => codec.writeToken(resultCodec, SonNumber(CS_LONG))
+
+    case D_BOOLEAN => codec.writeToken(resultCodec, SonBoolean(CS_BOOLEAN))
+
+    case D_NULL => ??? //TODO what do we do in this case
+  }
+
+  /**
+    *
+    * @param dataType
+    * @param codec
+    * @param currentResCodec
+    * @return
+    */
+  private def processTypesArray(dataType: Int, codec: Codec, currentResCodec: Codec): Codec = {
+    dataType match {
+      case D_ZERO_BYTE =>
+        currentResCodec
+      case D_FLOAT_DOUBLE =>
+        codec.writeToken(currentResCodec, codec.readToken(SonNumber(CS_DOUBLE)))
+      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
+        codec.writeToken(currentResCodec, codec.readToken(SonString(CS_STRING)))
+      case D_BSONOBJECT =>
+        codec.writeToken(currentResCodec, codec.readToken(SonObject(CS_OBJECT)))
+      case D_BSONARRAY =>
+        codec.writeToken(currentResCodec, codec.readToken(SonArray(CS_ARRAY)))
+      case D_NULL => throw new Exception
+      case D_INT =>
+        codec.writeToken(currentResCodec, codec.readToken(SonNumber(CS_INTEGER)))
+      case D_LONG =>
+        codec.writeToken(currentResCodec, codec.readToken(SonNumber(CS_LONG)))
+      case D_BOOLEAN =>
+        codec.writeToken(currentResCodec, codec.readToken(SonBoolean(CS_BOOLEAN)))
+    }
+  }
+
+  /**
+    *
+    * @param codec
+    * @param curentResCodec
+    * @param seqType
+    * @param f
+    * @tparam T
+    * @return
+    */
+  private def modifierAll[T](codec: Codec, curentResCodec: Codec, seqType: Int, f: T => T): Codec = {
+    seqType match {
+      case D_FLOAT_DOUBLE =>
+        val value0 = codec.readToken(SonNumber(CS_DOUBLE))
+        applyFunction(f, value0) match {
+          case value: SonNumber => codec.writeToken(curentResCodec, value)
+        }
+      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
+        val value0 = codec.readToken(SonString(CS_STRING))
+        applyFunction(f, value0) match {
+          case value: SonString => codec.writeToken(curentResCodec, value)
+        }
+      case D_BSONOBJECT =>
+        val value0 = codec.readToken(SonObject(CS_OBJECT))
+        applyFunction(f, value0) match {
+          case value: SonObject => codec.writeToken(curentResCodec, value)
+        }
+      case D_BSONARRAY =>
+        val value0 = codec.readToken(SonArray(CS_ARRAY))
+        applyFunction(f, value0) match {
+          case value: SonArray => codec.writeToken(curentResCodec, value)
+        }
+      case D_BOOLEAN =>
+        val value0 = codec.readToken(SonBoolean(CS_BOOLEAN))
+        applyFunction(f, value0) match {
+          case value: SonBoolean => codec.writeToken(curentResCodec, value)
+        }
+      case D_NULL =>
+        throw new Exception //TODO
+      case D_INT =>
+        val value0 = codec.readToken(SonNumber(CS_INTEGER))
+        applyFunction(f, value0) match {
+          case value: SonNumber => codec.writeToken(curentResCodec, value)
+        }
+      case D_LONG =>
+        val value0 = codec.readToken(SonNumber(CS_LONG))
+        applyFunction(f, value0) match {
+          case value: SonNumber => codec.writeToken(curentResCodec, value)
+        }
+    }
+  }
+
+  /**
+    *
+    * @param codec
+    * @param dataType
+    * @param injFunction
+    * @tparam T
+    * @return
+    */
+  private def modifierEnd[T](codec: Codec, dataType: Int, injFunction: T => T, codecRes: Codec, codecResCopy: Codec): (Codec, Codec) = dataType match {
+
+    case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
+      val token = codec.readToken(SonString(CS_STRING))
+      val value0 = token match {
+        case SonString(_, data) => data.asInstanceOf[String]
+      }
+      val resCodec = applyFunction(injFunction, value0) match {
+        case str: String => codec.writeToken(codecRes, SonString(CS_STRING, str))
+      }
+      val resCodecCopy = codec.writeToken(codecResCopy, token)
+      (resCodec, resCodecCopy)
+
+    case D_BSONOBJECT =>
+      val token = codec.readToken(SonObject(CS_OBJECT))
+      val resCodecCopy = codec.writeToken(codecResCopy, token)
+      token match {
+        case SonObject(_, data) => data match {
+          case byteBuf: ByteBuf =>
+            val resCodec = applyFunction(injFunction, byteBuf.array()) match {
+              case arr: Array[Byte] => codec.writeToken(codecRes, SonArray(CS_ARRAY, arr))
+            }
+            (resCodec, resCodecCopy)
+
+          case str: String =>
+            val resCodec = applyFunction(injFunction, str) match {
+              case resString: String => codec.writeToken(codecRes, SonString(CS_STRING, resString))
+            }
+            (resCodec, resCodecCopy)
+        }
+      }
+
+    case D_BSONARRAY =>
+      val token = codec.readToken(SonArray(CS_ARRAY))
+      val resCodecCopy = codec.writeToken(codecResCopy, token)
+      token match {
+        case SonArray(_, data) => data match {
+          case byteArr: Array[Byte] =>
+            val resCodec = applyFunction(injFunction, byteArr) match {
+              case resByteArr: Array[Byte] => codec.writeToken(codecRes, SonArray(CS_ARRAY, resByteArr))
+            }
+            (resCodec, resCodecCopy)
+
+          case jsonString: String =>
+            val resCodec = applyFunction(injFunction, jsonString) match {
+              case resString: String => codec.writeToken(codecRes, SonString(CS_STRING, resString))
+            }
+            (resCodec, resCodecCopy)
+        }
+      }
+
+
+    case D_BOOLEAN =>
+      val token = codec.readToken(SonBoolean(CS_BOOLEAN))
+      val value0: Boolean = token match {
+        case SonBoolean(_, data) => data.asInstanceOf[Boolean]
+      }
+      val resCodec = applyFunction(injFunction, value0) match {
+        case tkn: Boolean => codec.writeToken(codecRes, SonBoolean(CS_BOOLEAN, tkn))
+      }
+      val resCodecCopy = codec.writeToken(codecResCopy, token)
+      (resCodec, resCodecCopy)
+
+    case D_FLOAT_DOUBLE =>
+      val token = codec.readToken(SonNumber(CS_DOUBLE))
+      val value0: Double = token match {
+        case SonNumber(_, data) => data.asInstanceOf[Double]
+      }
+      val resCodec = applyFunction(injFunction, value0) match {
+        case tkn: Double => codec.writeToken(codecRes, SonNumber(CS_DOUBLE, tkn))
+      }
+      val resCodecCopy = codec.writeToken(codecResCopy, token)
+      (resCodec, resCodecCopy)
+
+    case D_INT =>
+      val token = codec.readToken(SonNumber(CS_INTEGER))
+      val value0: Int = token match {
+        case SonNumber(_, data) => data.asInstanceOf[Int]
+      }
+      val resCodec = applyFunction(injFunction, value0) match {
+        case tkn: Int => codec.writeToken(codecRes, SonNumber(CS_INTEGER, tkn))
+      }
+      val resCodecCopy = codec.writeToken(codecResCopy, token)
+      (resCodec, resCodecCopy)
+
+    case D_LONG =>
+
+      val token = codec.readToken(SonNumber(CS_LONG))
+      val value0: Long = token match {
+        case SonNumber(_, data) => data.asInstanceOf[Long]
+      }
+      val resCodec = applyFunction(injFunction, value0) match {
+        case tkn: Long => codec.writeToken(codecRes, SonNumber(CS_LONG, tkn))
+      }
+      val resCodecCopy = codec.writeToken(codecResCopy, token)
+      (resCodec, resCodecCopy)
+
+    case D_NULL => throw new Exception() //TODO Implement
+
+
+  }
+
+  /**
+    *
+    * @param statementsList
+    * @param seqType
+    * @param codec
+    * @param currentResCodec
+    * @param fieldID
+    * @param injFunction
+    * @tparam T
+    * @return
+    */
+  private def processTypesAll[T](statementsList: StatementsList, seqType: Int, codec: Codec, currentResCodec: Codec, fieldID: String, injFunction: T => T): Codec = {
+    seqType match {
+      case D_FLOAT_DOUBLE =>
+        codec.writeToken(currentResCodec, codec.readToken(SonNumber(CS_DOUBLE)))
+      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
+        codec.writeToken(currentResCodec, codec.readToken(SonString(CS_STRING)))
+      case D_BSONOBJECT =>
+        val partialCodec: Codec = codec.readToken(SonObject(CS_OBJECT)) match {
+          case SonObject(_, result) => result match {
+            case byteBuf: ByteBuf => CodecObject.toCodec(byteBuf)
+            case jsonString: String => CodecObject.toCodec(jsonString)
+          }
+        }
+        val codecAux = modifyAll(statementsList, partialCodec, fieldID, injFunction)
+        currentResCodec + codecAux
+      case D_BSONARRAY =>
+        val partialCodec: Codec = codec.readToken(SonArray(CS_ARRAY)) match {
+          case SonArray(_, result) => result match {
+            case byteBuf: ByteBuf => CodecObject.toCodec(byteBuf)
+            case jsonString: String => CodecObject.toCodec(jsonString)
+          }
+        }
+        val codecAux = modifyAll(statementsList, partialCodec, fieldID, injFunction)
+        currentResCodec + codecAux
+      case D_NULL => throw new Exception
+      case D_INT =>
+        codec.writeToken(currentResCodec, codec.readToken(SonNumber(CS_INTEGER)))
+      case D_LONG =>
+        codec.writeToken(currentResCodec, codec.readToken(SonNumber(CS_LONG)))
+      case D_BOOLEAN =>
+        codec.writeToken(currentResCodec, codec.readToken(SonBoolean(CS_BOOLEAN)))
+    }
+  }
+
   /**
     * Verifies if Key given by user is HalfWord and if it matches with the one extracted.
     *
@@ -1067,7 +1730,7 @@ class BosonImpl(
     * @param extracted Key extracted.
     * @return
     */
-  def isHalfword(fieldID: String, extracted: String): Boolean = {
+  private def isHalfword(fieldID: String, extracted: String): Boolean = {
     if (fieldID.contains(STAR) & extracted.nonEmpty) {
       val list: Array[String] = fieldID.split(STAR_CHAR)
       (extracted, list.length) match {
@@ -1100,1438 +1763,373 @@ class BosonImpl(
   }
 
   /**
-    * Function used to apply the final function given bbyb the used
+    * Method that tries to apply the given injector function to a given value
     *
-    * @param f     Fucntion to be applied when the final value is extracted or when the position to inject has been found
-    * @param value Value on which the function is applied to
-    * @tparam T Type of the value to be extracted of injected
-    * @return Returns the value resulting from applying the value to the function
+    * @param injFunction - The injector function to be applied
+    * @param value       - The value to apply the injector function to
+    * @tparam T - The type of the value
+    * @return A modified value in which the injector function was applied
     */
-  private def applyFunction[T](f: T => T, value: Any): T = {
-    Try(f(value.asInstanceOf[T])) match {
-      case Success(v) =>
-        v.asInstanceOf[T]
-      case Failure(e) =>
-        value match {
-          case x: Double =>
-            Try(f(x.toFloat.asInstanceOf[T])) match {
-              case Success(v) =>
-                v.asInstanceOf[T]
-              case Failure(e1) =>
-                throw CustomException(s"Type Error. Cannot Cast ${value.getClass.getSimpleName.toLowerCase} inside the Injector Function.")
-            }
-          case x: Array[Byte] =>
-            Try(f(new String(x).asInstanceOf[T])) match {
-              case Success(v) =>
-                v.asInstanceOf[T]
-              case Failure(e1) =>
-                Try(f(Instant.parse(new String(x)).asInstanceOf[T])) match {
-                  case Success(v) =>
-                    v.asInstanceOf[T]
-                  case Failure(e2) =>
-                    throw CustomException(s"Type Error. Cannot Cast ${value.getClass.getSimpleName.toLowerCase} inside the Injector Function.")
-                }
-            }
+  private def applyFunction[T](injFunction: T => T, value: Any): T = {
+
+    Try(injFunction(value.asInstanceOf[T])) match {
+      case Success(modifiedValue) =>
+        println("Ele entrou aqui efetivamente")
+        modifiedValue
+
+      case Failure(_) => value match {
+        case double: Double =>
+          Try(injFunction(double.toFloat.asInstanceOf[T])) match { //try with the value being a Double
+            case Success(modifiedValue) => modifiedValue.asInstanceOf[T]
+
+            case Failure(_) => throw CustomException(s"Type Error. Cannot Cast ${value.getClass.getSimpleName.toLowerCase} inside the Injector Function.")
+          }
+
+        case byteArr: Array[Byte] =>
+          println("ENTROU AQUIII")
+          Try(injFunction(new String(byteArr).asInstanceOf[T])) match { //try with the value being a Array[Byte]
+            case Success(modifiedValue) =>
+              println("foi succ?")
+              println(modifiedValue.asInstanceOf[T])
+              println("da treta")
+              modifiedValue.asInstanceOf[T]
+
+            case Failure(_) =>
+              Try(injFunction(Instant.parse(new String(byteArr)).asInstanceOf[T])) match { //try with the value being an Instant
+                case Success(modifiedValue) => modifiedValue.asInstanceOf[T]
+
+                case Failure(_) => throw CustomException(s"Type Error. Cannot Cast ${value.getClass.getSimpleName.toLowerCase} inside the Injector Function.")
+              }
+          }
+      }
+
+      case _ => throw CustomException(s"Type Error. Cannot Cast ${value.getClass.getSimpleName.toLowerCase} inside the Injector Function.")
+    }
+  }
+
+  /**
+    *
+    * @param statementsList
+    * @param codec
+    * @param currentCodec
+    * @param injFunction
+    * @param key
+    * @param left
+    * @param mid
+    * @param right
+    * @tparam T
+    * @return
+    */
+  def arrayInjection[T](statementsList: StatementsList, codec: Codec, currentCodec: Codec, injFunction: T => T, key: String, left: Int, mid: String, right: Any): Codec = {
+    val arrayTokenCodec = codec.readToken(SonArray(CS_ARRAY)) match {
+      case SonArray(_, data) => data match {
+        case byteBuf: ByteBuf => CodecObject.toCodec(byteBuf)
+        case jsonString: String => CodecObject.toCodec(jsonString)
+      }
+    }
+    val codecArrayEnd: Codec = (key, left, mid.toLowerCase(), right) match {
+      case (EMPTY_KEY, from, expr, to) if to.isInstanceOf[Int] =>
+        modifyArrayEnd(statementsList, arrayTokenCodec, injFunction, expr, from.toString, to.toString)
+      case (EMPTY_KEY, from, expr, _) =>
+        modifyArrayEnd(statementsList, arrayTokenCodec, injFunction, expr, from.toString)
+      case (nonEmptyKey, from, expr, to) if to.isInstanceOf[Int] =>
+        modifyArrayEndWithKey(statementsList, arrayTokenCodec, nonEmptyKey, injFunction, expr, from.toString, to.toString)
+      case (nonEmptyKey, from, expr, _) =>
+        modifyArrayEndWithKey(statementsList, arrayTokenCodec, nonEmptyKey, injFunction, expr, from.toString)
+    }
+    currentCodec + codecArrayEnd
+  }
+
+  /**
+    *
+    * @param statementsList
+    * @param codec
+    * @param injFunction
+    * @param condition
+    * @param from
+    * @param to
+    * @tparam T
+    * @return
+    */
+  def modifyArrayEnd[T](statementsList: StatementsList, codec: Codec, injFunction: T => T, condition: String, from: String, to: String = C_END): Codec = {
+    val startReaderIndex = codec.getReaderIndex
+    val originalSize = codec.readSize
+
+    def iterateDataStructure(currentCodec: Codec): Codec = {
+      if ((codec.getReaderIndex - startReaderIndex) < originalSize) //exceptions
+        currentCodec
+      else {
+        val dataType: Int = codec.readDataType
+        val codecWithDataType = codec.writeToken(currentCodec, SonNumber(CS_INTEGER, dataType))
+        dataType match {
+          case 0 => iterateDataStructure(codecWithDataType)
           case _ =>
-            throw CustomException(s"Type Error. Cannot Cast ${value.getClass.getSimpleName.toLowerCase} inside the Injector Function.")
-        }
-    }
-  }
-
-  // Injector Starts here
-
-  /**
-    * Function used to search for keys of interest recursively
-    *
-    * @param list    A list with pairs that contains the key of interest and the type of operation
-    * @param codec   The codec abstraction which contains the structure to work on
-    * @param fieldID Key of interest
-    * @param f       Function used to inject a new value in the structure
-    * @tparam T Type of value to be injected
-    * @return ByteBuf containing the new value injected
-    */
-  @deprecated
-  def modifyAll[T](list: List[(Statement, String)], codec: Codec, fieldID: String, f: T => T): Codec = {
-    /*
-    * If fieldID is empty, it returns the original ByteBuf
-    */
-    val startReader: Int = codec.getReaderIndex
-    val originalSize: Int = codec.readSize
-    while ((codec.getReaderIndex - startReader) < originalSize) {
-      val dataType: Int = codec.readDataType
-      dataType match {
-        case 0 => // Final
-          codec.writeToken(SonNumber(CS_BYTE, dataType))
-        case _ =>
-          val codecWithDataType = codec.writeToken(SonNumber(CS_BYTE, dataType))
-          val (key, b): (String, Int) = {
-            val key: String = codecWithDataType.readToken(SonString(CS_NAME)) match {
-              case SonString(_, keyString) => keyString.asInstanceOf[String]
-            }
-            val token = codecWithDataType.readToken(SonBoolean(C_ZERO))
-            val b: Byte = token match {
-              case SonBoolean(_, result) => result.asInstanceOf[Byte]
-            }
-            (key, b)
-          }
-          val codecWithKey = codecWithDataType.writeToken(SonString(CS_STRING, key)).writeToken(SonNumber(CS_BYTE, b))
-
-          key match {
-            case extracted if fieldID.toCharArray.deep == extracted.toCharArray.deep || isHalfword(fieldID, extracted) =>
-              /*
-               * Found a field equal to the key.
-               * Perform injection
-               */
-              if (list.lengthCompare(1) == 0) { //If the size of the list is 1
-                if (list.head._2.contains(C_DOUBLEDOT)) {
-                  dataType match {
-                    case D_BSONOBJECT | D_BSONARRAY =>
-                      val token = if (dataType == D_BSONOBJECT) SonObject(CS_OBJECT) else SonArray(CS_ARRAY)
-                      val partialData: Either[ByteBuf, String] = codecWithKey.readToken(token) match {
-                        case SonObject(_, result) => result match {
-                          case byteBuf: ByteBuf => Left(byteBuf)
-                          case jsonString: String => Right(jsonString)
-                        }
-                        case SonArray(_, result) => result match {
-                          case byteBuf: ByteBuf => Left(byteBuf)
-                          case jsonString: String => Right(jsonString)
-                        }
-                      }
-                      val modifiedCodec = execStatementPatternMatch(partialData, list, f)
-                      modifierAll(modifiedCodec, dataType, f)
-
-                    case _ =>
-                      modifierAll(codecWithKey, dataType, f)
-                  }
-                } else modifierAll(codecWithKey, dataType, f)
-
-              } else {
-                if (list.head._2.contains(C_DOUBLEDOT)) {
-                  dataType match {
-                    case (D_BSONOBJECT | D_BSONARRAY) =>
-                      val token = if (dataType == D_BSONOBJECT) SonObject(CS_OBJECT) else SonArray(CS_ARRAY)
-                      val partialData: Either[ByteBuf, String] = codecWithKey.readToken(token) match {
-                        case SonObject(_, result) => result match {
-                          case byteBuf: ByteBuf => Left(byteBuf)
-                          case jsonString: String => Right(jsonString)
-                        }
-                        case SonArray(_, result) => result match {
-                          case byteBuf: ByteBuf => Left(byteBuf)
-                          case jsonString: String => Right(jsonString)
-                        }
-                      }
-                      val modifiedCodec: Codec = execStatementPatternMatch(partialData, list.drop(1), f)
-                      execStatementPatternMatch(modifiedCodec.getCodecData, list, f)
-                    //                      buf2.release()
-                    //                      result.writeBytes(buf3)
-                    //                      buf3.release()
-                    case _ => processTypesAll(list, dataType, codecWithKey, fieldID, f)
-                  }
-                } else {
-                  dataType match {
-                    case (D_BSONOBJECT | D_BSONARRAY) =>
-                      execStatementPatternMatch(codecWithKey.getCodecData, list.drop(1), f)
-                    case _ =>
-                      processTypesArray(dataType, codecWithKey)
-                  }
-                }
+            val (isArray, key, byte): (Boolean, String, Int) = {
+              val key: String = codec.readToken(SonString(CS_NAME)) match {
+                case SonString(_, keyString) => keyString.asInstanceOf[String]
               }
-            case x if fieldID.toCharArray.deep != x.toCharArray.deep && !isHalfword(fieldID, x) =>
-              /*
-              * Didn't find a field equal to key
-              * Consume value and check deeper Levels
-              * */
-              if (list.head._2.contains(C_DOUBLEDOT))
-                processTypesAll(list, dataType, codecWithKey, fieldID, f)
-              else
-                processTypesArray(dataType, codecWithKey)
-          }
-          codecWithKey
-      }
-    }
-    codec //TODO Maybe change this ? is this correct
-    //    result.capacity(result.\ writerIndex())
-    //    val finalResult: ByteBuf = Unpooled.buffer(result.capacity() + 4).writeIntLE(result.capacity() + 4).writeBytes(result)
-    //    result.release()
-    //    finalResult
-  }
-
-  /**
-    * Function used to copied the values inside arrays that aren´t of interest to the resulting structure
-    *
-    * @param dataType The type on the value we are reading and writting on the new structure
-    * @param codec    Structure from where we read the values
-    * @return Doesn't return anything because the changes are reflected in the result structure
-    */
-  @deprecated
-  private def processTypesArray(dataType: Int, codec: Codec): Codec = {
-    dataType match {
-      case D_ZERO_BYTE => throw new Exception //TODO - NOT here!!
-      case D_FLOAT_DOUBLE =>
-        codec.writeToken(codec.readToken(SonNumber(CS_DOUBLE)))
-      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
-        codec.writeToken(codec.readToken(SonString(CS_STRING)))
-      case D_BSONOBJECT =>
-        codec.writeToken(codec.readToken(SonObject(CS_OBJECT))) // TODO - codec + codecAux ???
-      case D_BSONARRAY =>
-        codec.writeToken(codec.readToken(SonArray(CS_ARRAY))) // TODO - codec + codecAux ???
-      case D_NULL => throw Exception
-      case D_INT =>
-        codec.writeToken(codec.readToken(SonNumber(CS_INTEGER)))
-      case D_LONG =>
-        codec.writeToken(codec.readToken(SonNumber(CS_LONG)))
-      case D_BOOLEAN =>
-        codec.writeToken(codec.readToken(SonBoolean(CS_BOOLEAN)))
-    }
-  }
-
-  /**
-    * Function used to perform injection of the new values
-    *
-    * @param codec   Structure from which we are reading the old values
-    * @param seqType Type of the value found and processing
-    * @param f       Function given by the user with the new value
-    * @tparam T Type of the value being injected
-    */
-  @deprecated
-  private def modifierAll[T](codec: Codec, seqType: Int, f: T => T): Codec = {
-    seqType match {
-      case D_FLOAT_DOUBLE =>
-        val value0 = codec.readToken(SonNumber(CS_DOUBLE))
-        applyFunction(f, value0) match {
-          case value: SonNumber => codec.writeToken(value)
-        }
-      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
-        val value0 = codec.readToken(SonString(CS_STRING))
-        applyFunction(f, value0) match {
-          case value: SonString => codec.writeToken(value)
-        }
-      case D_BSONOBJECT =>
-        val value0 = codec.readToken(SonObject(CS_OBJECT))
-        applyFunction(f, value0) match {
-          case value: SonObject => codec.writeToken(value) // TODO - codec + codecAux ???
-        }
-      case D_BSONARRAY =>
-        val value0 = codec.readToken(SonArray(CS_ARRAY))
-        applyFunction(f, value0) match {
-          case value: SonArray => codec.writeToken(value) // TODO - codec + codecAux ???
-        }
-      case D_BOOLEAN =>
-        val value0 = codec.readToken(SonBoolean(CS_BOOLEAN))
-        applyFunction(f, value0) match {
-          case value: SonBoolean => codec.writeToken(value)
-        }
-      case D_NULL =>
-        throw new Exception //TODO
-      case D_INT =>
-        val value0 = codec.readToken(SonNumber(CS_INTEGER))
-        applyFunction(f, value0) match {
-          case value: SonNumber => codec.writeToken(value)
-        }
-      case D_LONG =>
-        val value0 = codec.readToken(SonNumber(CS_LONG))
-        applyFunction(f, value0) match {
-          case value: SonNumber => codec.writeToken(value)
-        }
-    }
-  }
-
-  /**
-    * Function used to copy the values inside objects that aren't of interest to the resulting structure
-    *
-    * @param list    A list with pairs that contains the key of interest and the type of operation
-    * @param seqType Type of the value found and processing
-    * @param codec   Structure from which we are reading the old values
-    * @param fieldID name of the field we are searching
-    * @param f       Function given by the user with the new value
-    * @tparam T Type of the value being injected
-    */
-  @deprecated
-  private def processTypesAll[T](list: List[(Statement, String)], seqType: Int, codec: Codec, fieldID: String, f: T => T): Codec = {
-    seqType match {
-      case D_FLOAT_DOUBLE =>
-        codec.writeToken(codec.readToken(SonNumber(CS_DOUBLE)))
-      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
-        codec.writeToken(codec.readToken(SonString(CS_STRING)))
-      case D_BSONOBJECT =>
-        val partialCodec: Codec = codec.readToken(SonObject(CS_OBJECT)) match {
-          case SonObject(_, result) => CodecObject.toCodec(result)
-        }
-        val codecAux = modifyAll(list, partialCodec, fieldID, f)
-        // TODO - codec + codecAux ???
-        codec.writeToken(codecAux.readToken(SonObject(CS_OBJECT)))
-      case D_BSONARRAY =>
-        val partialCodec: Codec = codec.readToken(SonArray(CS_ARRAY)) match {
-          case SonArray(_, result) => CodecObject.toCodec(result)
-        }
-        val codecAux = modifyAll(list, partialCodec, fieldID, f)
-        // TODO - codec + codecAux ???
-        codec.writeToken(codecAux.readToken(SonArray(CS_ARRAY)))
-      case D_NULL => throw Exception
-      case D_INT =>
-        codec.writeToken(codec.readToken(SonNumber(CS_INTEGER)))
-      case D_LONG =>
-        codec.writeToken(codec.readToken(SonNumber(CS_LONG)))
-      case D_BOOLEAN =>
-        codec.writeToken(codec.readToken(SonBoolean(CS_BOOLEAN)))
-    }
-  }
-
-  /**
-    * Function used to perform the injection on the last ocurrence of a field
-    *
-    * @param codec   Structure from which we are reading the old values
-    * @param seqType Type of the value found and processing
-    * @param f       Function given by the user with the new value
-    * @tparam T Type of the value being injected
-    */
-  @deprecated
-  private def modifierEnd[T](codec: Codec, seqType: Int, f: T => T): (Codec, Codec) = {
-    seqType match {
-      case D_FLOAT_DOUBLE =>
-        val value0 = codec.readToken(SonNumber(CS_DOUBLE))
-        val codecCopy = codec.writeToken(value0)
-        val codecRes = applyFunction(f, value0) match {
-          case x: SonNumber => codec.writeToken(x)
-        }
-        (codecRes, codecCopy)
-      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
-        val value0 = codec.readToken(SonString(CS_STRING))
-        val codecCopy = codec.writeToken(value0)
-        val codeRes = applyFunction(f, value0) match {
-          case value: SonString => codec.writeToken(value)
-        }
-        (codeRes, codecCopy)
-      case D_BSONOBJECT =>
-        val value0 = codec.readToken(SonObject(CS_OBJECT))
-        val codecCopy = codec.writeToken(value0)
-        val codecRes = applyFunction(f, value0) match {
-          case value: SonObject => codec.writeToken(value) // TODO - codec + codecAux ???
-        }
-        (codecRes, codecCopy)
-      case D_BSONARRAY =>
-        val value0 = codec.readToken(SonArray(CS_ARRAY))
-        val codecCopy = codec.writeToken(value0)
-        val codecRes = applyFunction(f, value0) match {
-          case value: SonArray => codec.writeToken(value) // TODO - codec + codecAux ???
-        }
-        (codecRes, codecCopy)
-      case D_BOOLEAN =>
-        val value0 = codec.readToken(SonBoolean(CS_BOOLEAN))
-        val codecCopy = codec.writeToken(value0)
-        val codecRes = applyFunction(f, value0) match {
-          case value: SonBoolean => codec.writeToken(value)
-        }
-        (codecRes, codecCopy)
-      case D_NULL =>
-        throw new Exception //TODO - Change
-      case D_INT =>
-        val value0 = codec.readToken(SonNumber(CS_INTEGER))
-        val codecCopy = codec.writeToken(value0)
-        val codecRes = applyFunction(f, value0) match {
-          case value: SonNumber => codec.writeToken(value)
-        }
-        (codecRes, codecCopy)
-      case D_LONG =>
-        val value0 = codec.readToken(SonNumber(CS_LONG))
-        val codecCopy = codec.writeToken(value0)
-        val codecRes = applyFunction(f, value0) match {
-          case value: SonNumber => codec.writeToken(value)
-        }
-        (codecRes, codecCopy)
-    }
-  }
-
-  /**
-    * Function used to perform the search on the last element of an array
-    *
-    * @param list      A list with pairs that contains the key of interest and the type of operation
-    * @param codec     Structure from which we are reading the old values
-    * @param f         Function given by the user with the new value
-    * @param condition Represents a type of injection, it can be END, ALL, FIRST, # TO #, # UNTIL #
-    * @param limitInf  Represent the inferior limit when a range is given
-    * @param limitSup  Represent the superior limit when a range is given
-    * @tparam T Type of the value being injected
-    * @return A new bosonImpl with the new values injected
-    */
-  @deprecated
-  def modifyArrayEnd[T](list: List[(Statement, String)], codec: Codec, f: T => T, condition: String, limitInf: String = C_ZERO, limitSup: String = C_END): BosonImpl = {
-    /*
-    * If fieldID is empty returns the Original BosomImpl
-    * */
-    val startReaderIndex: Int = codec.getReaderIndex
-    val originalSize: Int = codec.readSize
-    val exceptions: ListBuffer[Throwable] = new ListBuffer[Throwable]
-    while ((codec.getReaderIndex - startReaderIndex) < originalSize && exceptions.size < 2) {
-      val dataType: Int = codec.readDataType
-      //      result.writeByte(dataType)
-      //      resultCopy.writeByte(dataType)
-      dataType match {
-        case 0 => codec.writeToken(SonNumber(CS_BYTE, dataType))
-        case _ =>
-          val codecWithDataType = codec.writeToken(SonNumber(CS_BYTE, dataType))
-          val (isArray, key, b): (Boolean, String, Int) = {
-            val key: String = codecWithDataType.readToken(SonString(CS_NAME)) match {
-              case SonString(_, keyString) => keyString.asInstanceOf[String]
+              val byte: Byte = codec.readToken(SonBoolean(C_ZERO)) match {
+                case SonBoolean(_, result) => result.asInstanceOf[Byte]
+              }
+              (key.forall(b => b.isDigit), key, byte)
             }
-            //            val key: ListBuffer[Byte] = new ListBuffer[Byte]
-            //            while (codec.getByte(codec.getReaderIndex) != 0 || key.length < 1) {
-            //              val b: Byte = codec.readByte()
-            //              key.append(b)
-            //            }
-            val token = codecWithDataType.readToken(SonBoolean(C_ZERO))
-            val b: Byte = token match {
-              case SonBoolean(_, result) => result.asInstanceOf[Byte]
-            }
-            (key.forall(byte => byte.isDigit), key, b)
-          }
-          val codecWithKey = codecWithDataType.writeToken(SonString(CS_STRING, key))
-          //          result.writeBytes(key).writeByte(b)
-          //          resultCopy.writeBytes(key).writeByte(b)
-          val keyString: String = new String(key)
-          (keyString, condition, limitSup) match {
-            case (x, C_END, _) if isArray =>
-              exceptions.clear()
-              if (list.size == 1) {
-                if (list.head._2.contains(C_DOUBLEDOT)) {
-                  dataType match {
-                    case (D_BSONOBJECT | D_BSONARRAY) =>
-                      if (exceptions.isEmpty) {
-                        //                        val size: Int = codec.readSize
-                        //TODO - Create function to read Obj or Arr (Maybe) +
-                        val partialCodec: Either[ByteBuf, String] = codec.readToken(SonArray(CS_ARRAY)) match {
-                          case SonArray(_, value) => value.asInstanceOf[Either[ByteBuf, String]]
-                          case SonObject(_, value) => value.asInstanceOf[Either[ByteBuf, String]]
-                        }
-                        //                        val bufRes: ByteBuf = Unpooled.buffer()
-                        //                        val bufResCopy: ByteBuf = Unpooled.buffer()
-                        //                        result.clear().writeBytes(resultCopy.duplicate()) //TODO - Why!?!
-                        val codec3: Codec = {
-                          //TODO - Change names when I understand what they are
-                          if (list.head._1.isInstanceOf[ArrExpr])
-                            execStatementPatternMatch(partialCodec, list, f)
-                          else
-                            CodecObject.toCodec(partialCodec)
+            val codecWithKey = codec.writeToken(codecWithDataType, SonString(CS_STRING, key))
+
+            (new String(key), condition, to) match {
+              case (x, C_END, _) if isArray =>
+                //exception = 0
+                if (statementsList.size == 1) {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT)) {
+                    dataType match {
+                      case D_BSONOBJECT | D_BSONARRAY =>
+                        val partialData: Either[ByteBuf, String] = codec.readToken(SonArray(CS_ARRAY)) match {
+                          case SonArray(_, value) => value match {
+                            case byteBuf: ByteBuf => Left(byteBuf)
+                            case jsonString: String => Right(jsonString)
+                          }
                         }
 
-                        Try(modifierEnd(codec3, dataType, f)) match {
-                          case Success(codecTuple) => codecTuple
-                          case Failure(e) =>
-                            exceptions.append(e)
+                        val partialCodec = {
+                          if (statementsList.head._1.isInstanceOf[ArrExpr])
+                            inject(partialData, statementsList, injFunction)
+                          else {
+                            partialData match {
+                              case Left(byteBuf) => CodecObject.toCodec(byteBuf)
+                              case Right(jsonString) => CodecObject.toCodec(jsonString)
+                            }
+                          }
                         }
-                        //                        (codecRes,codecResCopy)
-                        //                        result.writeBytes(bufRes)
-                        //                        resultCopy.writeBytes(bufResCopy)
-                        //                        buf1.release()
-                        //                        bufRes.release()
-                        //                        bufResCopy.release()
-                        //                        buf3.release()
-                      }
-                    case _ =>
-                      if (exceptions.isEmpty) {
-                        //                        result.clear().writeBytes(resultCopy.duplicate())
-                        Try(modifierEnd(codec, dataType, f)) match {
-                          case Success(codecTuple) => codecTuple
-                          case Failure(e) =>
-                            exceptions.append(e)
-                        }
-                      }
+
+                        //modifierEnd(partialCodec, dataType, injFunction)
+                        ???
+                      case _ =>
+                        //modifier(codec, dataType, injFunction)
+                        ???
+                    }
+                  } else {
+                    //modifierEnd(codec, dataType, injFunction)
+                    ???
                   }
                 } else {
-                  if (exceptions.isEmpty) {
-                    //                    result.clear().writeBytes(resultCopy.duplicate())
-                    Try(modifierEnd(codec, dataType, f)) match {
-                      case Success(codecTuple) => codecTuple
-                      case Failure(e) =>
-                        exceptions.append(e)
-                    }
-                  }
-                }
-              } else {
-                if (list.head._2.contains(C_DOUBLEDOT) /*&& list.head._1.isInstanceOf[ArrExpr]*/ ) {
-                  dataType match {
-                    case (D_BSONARRAY | D_BSONOBJECT) =>
-                      if (exceptions.isEmpty) {
-                        //                        result.clear().writeBytes(resultCopy.duplicate())
-                        //                        val size: Int = codec.getIntLE(codec.readerIndex())
-                        //                        val buf1: ByteBuf = codec.readBytes(size)
-                        val codec1 = codec.readToken(SonArray(CS_ARRAY)) match {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT) && statementsList.head._1.isInstanceOf[ArrExpr]) {
+                    dataType match {
+                      case D_BSONARRAY | D_BSONOBJECT =>
+                        val partialData: Either[ByteBuf, String] = codec.readToken(SonArray(CS_ARRAY)) match {
                           case SonArray(_, value) => value.asInstanceOf[Either[ByteBuf, String]]
                         }
-                        // val buf2: ByteBuf = execStatementPatternMatch(buf1.duplicate(), list, f)
-                        val buf2: Either[ByteBuf, String] =
-                          if (list.head._1.isInstanceOf[ArrExpr])
-                            execStatementPatternMatch(codec1, list, f).getCodecData //I DON'T KNOW!!!!
+
+                        val auxCodec: Codec = inject(partialData, statementsList, injFunction)
+
+                        val partialCodec = {
+                          if (statementsList.head._1.isInstanceOf[ArrExpr])
+                            inject(auxCodec.getCodecData, statementsList.drop(1), injFunction)
                           else
-                            codec1
-                        //                            Unpooled.buffer(buf1.capacity()).writeBytes(buf1)
-                        //val buf3: ByteBuf =
-                        Try(execStatementPatternMatch(buf2, list.drop(1), f)) match {
-                          case Success(v) =>
-                            codec.writeToken(SonNumber(CS_BYTE, v)) // Not right, v is a Codec
-                          //                            result.writeBytes(v)
-                          //                            v.release()
-                          //                            resultCopy.writeBytes(buf2)
-                          case Failure(e) =>
-                            result.writeBytes(buf2.duplicate())
-                            resultCopy.writeBytes(buf2)
-                            exceptions.append(e)
+                            inject(partialData, statementsList.drop(1), injFunction)
                         }
-
-                        //processTypesArray(dataType, buffer, resultCopy)
-                        buf1.release()
-                        buf2.release()
-                        //buf3.release()
-                      }
-                    case _ =>
-                      processTypesArrayEnd(list, EMPTY_KEY, dataType, codec, f, condition, limitInf, limitSup, result, resultCopy)
-                  }
-                } else {
-                  dataType match {
-                    case (D_BSONARRAY | D_BSONOBJECT) =>
-                      if (exceptions.isEmpty) {
-                        result.clear().writeBytes(resultCopy.duplicate())
-                        // val res: ByteBuf =
-                        Try(execStatementPatternMatch(codec.duplicate.getCodecData, list.drop(1), f)) match {
-                          case Success(v) =>
-                            result.writeBytes(v)
-                            v.release()
-                            processTypesArray(dataType, codec, resultCopy)
-                          case Failure(e) =>
-                            processTypesArray(dataType, codec.duplicate(), result)
-                            processTypesArray(dataType, codec, resultCopy)
-                            exceptions.append(e)
-                        }
-                      }
-                    case _ =>
-                      processTypesArray(dataType, codec.duplicate(), result)
-                      processTypesArray(dataType, codec, resultCopy)
-                  }
-                }
-              }
-            case (x, _, C_END) if isArray && limitInf.toInt <= keyString.toInt =>
-              if (list.size == 1) {
-                if (list.head._2.contains(C_DOUBLEDOT) /*&& !list.head._1.isInstanceOf[KeyWithArrExpr]*/ ) {
-                  dataType match {
-                    case (D_BSONOBJECT | D_BSONARRAY) =>
-                      if (exceptions.isEmpty) {
-                        val size: Int = codec.getIntLE(codec.readerIndex())
-                        val buf1: ByteBuf = codec.readBytes(size)
-                        val bufRes: ByteBuf = Unpooled.buffer()
-                        val bufResCopy: ByteBuf = Unpooled.buffer()
-                        resultCopy.clear().writeBytes(result.duplicate())
-                        val buf3: ByteBuf = execStatementPatternMatch(buf1, list, f)
-                        Try(modifierEnd(buf3.duplicate(), dataType, f, bufRes, bufResCopy)) match {
-                          case Success(_) =>
-                          case Failure(e) =>
-                            exceptions.append(e)
-                        }
-                        result.writeBytes(bufRes)
-                        resultCopy.writeBytes(bufResCopy)
-                        buf1.release()
-                        bufRes.release()
-                        bufResCopy.release()
-                        buf3.release()
-                      } else
-                        exceptions.append(exceptions.head)
-                    case _ =>
-                      if (exceptions.isEmpty) {
-                        resultCopy.clear().writeBytes(result.duplicate())
-                        Try(modifierEnd(codec, dataType, f, result, resultCopy)) match {
-                          case Success(_) =>
-                          case Failure(e) =>
-                            exceptions.append(e)
-                        }
-                      } else
-                        exceptions.append(exceptions.head)
-                  }
-                } else {
-                  if (exceptions.isEmpty) {
-                    resultCopy.clear().writeBytes(result.duplicate())
-                    Try(modifierEnd(codec, dataType, f, result, resultCopy)) match {
-                      case Success(_) =>
-                      case Failure(e) =>
-                        exceptions.append(e)
+                      //currentCodec(Check) + partialCodec
+                      case _ =>
+                      //processTypesArrayEnd()
                     }
-                  } else
-                    exceptions.append(exceptions.head)
-                }
-              } else {
-                if (list.head._2.contains(C_DOUBLEDOT) /*&& list.head._1.isInstanceOf[ArrExpr]*/ ) {
-                  dataType match {
-                    case (D_BSONARRAY | D_BSONOBJECT) =>
-                      if (exceptions.isEmpty) {
-                        resultCopy.clear().writeBytes(result.duplicate())
-                        val size: Int = codec.getIntLE(codec.readerIndex())
-                        val buf1: ByteBuf = codec.duplicate().readRetainedSlice(size)
-                        val buf2: ByteBuf = execStatementPatternMatch(buf1, list, f)
-                        buf1.release()
-                        // val buf3: ByteBuf =
-                        Try(execStatementPatternMatch(buf2.duplicate(), list.drop(1), f)) match {
-                          case Success(v) =>
-                            buf2.release()
-                            result.writeBytes(v)
-                            v.release()
-                            processTypesArray(dataType, codec, resultCopy)
-                          case Failure(e) =>
-                            processTypesArray(dataType, codec.duplicate(), result)
-                            processTypesArray(dataType, codec, resultCopy)
-                            exceptions.append(e)
-                        }
-                      } else {
-                        exceptions.append(exceptions.head)
-                      }
-
-                    case _ =>
-                      processTypesArrayEnd(list, EMPTY_KEY, dataType, codec, f, condition, limitInf, limitSup, result, resultCopy)
-                  }
-                } else {
-                  dataType match {
-                    case (D_BSONARRAY | D_BSONOBJECT) =>
-                      if (exceptions.isEmpty) {
-                        resultCopy.clear().writeBytes(result.duplicate())
-                        //val res: ByteBuf =
-                        Try(execStatementPatternMatch(codec.duplicate(), list.drop(1), f)) match {
-                          case Success(v) =>
-                            result.writeBytes(v)
-                            processTypesArray(dataType, codec, resultCopy)
-                            v.release()
-                          case Failure(e) =>
-                            processTypesArray(dataType, codec.duplicate(), result)
-                            processTypesArray(dataType, codec, resultCopy)
-                            exceptions.append(e)
-                        }
-                      } else
-                        exceptions.append(exceptions.head)
-                    case _ =>
-                      processTypesArray(dataType, codec.duplicate(), result)
-                      processTypesArray(dataType, codec, resultCopy)
-                  }
-                }
-              }
-            case (x, _, C_END) if isArray && limitInf.toInt > keyString.toInt =>
-              if (list.head._2.contains(C_DOUBLEDOT) /*&& list.head._1.isInstanceOf[ArrExpr]*/ )
-                dataType match {
-                  case (D_BSONOBJECT | D_BSONARRAY) =>
-                    val size: Int = codec.getIntLE(codec.readerIndex())
-                    val buf1: ByteBuf = codec.readBytes(size)
-                    val buf2: ByteBuf = execStatementPatternMatch(buf1, list, f)
-                    result.writeBytes(buf2.duplicate())
-                    resultCopy.writeBytes(buf2.duplicate())
-                    buf1.release()
-                    buf2.release()
-                  case _ =>
-                    processTypesArray(dataType, codec.duplicate(), result)
-                    processTypesArray(dataType, codec, resultCopy)
-                }
-              else {
-                processTypesArray(dataType, codec.duplicate(), result)
-                processTypesArray(dataType, codec, resultCopy)
-              }
-            case (x, _, l) if isArray && (limitInf.toInt <= x.toInt && l.toInt >= x.toInt) =>
-              if (list.lengthCompare(1) == 0) {
-                if (list.head._2.contains(C_DOUBLEDOT)) {
-                  dataType match {
-                    case (D_BSONOBJECT | D_BSONARRAY) =>
-                      if (exceptions.isEmpty) {
-                        val size: Int = codec.getIntLE(codec.readerIndex())
-                        val buf1: ByteBuf = codec.readBytes(size)
-                        val bufRes: ByteBuf = Unpooled.buffer()
-                        val bufResCopy: ByteBuf = Unpooled.buffer()
-                        resultCopy.clear().writeBytes(result.duplicate())
-                        val buf3: ByteBuf = execStatementPatternMatch(buf1, list, f)
-                        Try(modifierEnd(buf3.duplicate(), dataType, f, bufRes, bufResCopy)) match {
-                          case Success(_) =>
-                          case Failure(e) =>
-                            exceptions.append(e)
-                        }
-                        result.writeBytes(bufRes)
-                        resultCopy.writeBytes(bufResCopy)
-                        buf1.release()
-                        bufRes.release()
-                        bufResCopy.release()
-                        buf3.release()
-                      } else
-                        exceptions.append(exceptions.head)
-                    case _ =>
-                      if (exceptions.isEmpty) {
-                        resultCopy.clear().writeBytes(result.duplicate())
-                        Try(modifierEnd(codec, dataType, f, result, resultCopy)) match {
-                          case Success(_) =>
-                          case Failure(e) =>
-                            exceptions.append(e)
-                        }
-                      } else
-                        exceptions.append(exceptions.head)
-
-                  }
-                } else {
-                  if (exceptions.isEmpty) {
-                    resultCopy.clear().writeBytes(result.duplicate())
-                    Try(modifierEnd(codec, dataType, f, result, resultCopy)) match {
-                      case Success(_) =>
-                      case Failure(e) =>
-                        exceptions.append(e)
+                  } else {
+                    dataType match { //TODO - I Don't understand this!!!!
+                      case D_BSONARRAY | D_BSONOBJECT =>
+                      case _ =>
                     }
-                  } else
-                    exceptions.append(exceptions.head)
+                  }
                 }
-              } else {
-                if (list.head._2.contains(C_DOUBLEDOT) /*&& list.head._1.isInstanceOf[ArrExpr]*/ ) {
-                  dataType match {
-                    case (D_BSONARRAY | D_BSONOBJECT) =>
-                      if (exceptions.isEmpty) {
-                        resultCopy.clear().writeBytes(result.duplicate())
-                        val size: Int = codec.getIntLE(codec.readerIndex())
-                        val buf1: ByteBuf = codec.duplicate().readBytes(size)
-                        val buf2: ByteBuf = execStatementPatternMatch(buf1, list, f)
-                        buf1.release()
-                        Try(execStatementPatternMatch(buf2.duplicate(), list.drop(1), f)) match {
-                          case Success(v) =>
-                            result.writeBytes(v)
-                            v.release()
-                            processTypesArray(dataType, codec, resultCopy)
-                          case Failure(e) =>
-                            processTypesArray(dataType, codec.duplicate(), result)
-                            processTypesArray(dataType, codec, resultCopy)
-                            exceptions.append(e)
-                        }
-                        buf2.release()
-                      } else
-                        exceptions.append(exceptions.head)
-                    case _ =>
-                      processTypesArrayEnd(list, EMPTY_KEY, dataType, codec, f, condition, limitInf, limitSup, result, resultCopy)
+              case (x, _, C_END) if isArray && from.toInt <= key.toInt =>
+                if (statementsList.size == 1) {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT) && !statementsList.head._1.isInstanceOf[KeyWithArrExpr]) {
+                    ???
+                  } else {
+                    ???
                   }
                 } else {
-                  dataType match {
-                    case (D_BSONARRAY | D_BSONOBJECT) =>
-                      if (exceptions.isEmpty) {
-                        resultCopy.clear().writeBytes(result.duplicate())
-                        Try(execStatementPatternMatch(codec.duplicate(), list.drop(1), f)) match {
-                          case Success(v) =>
-                            result.writeBytes(v)
-                            processTypesArray(dataType, codec, resultCopy)
-                            v.release()
-                          case Failure(e) =>
-                            processTypesArray(dataType, codec.duplicate(), result)
-                            processTypesArray(dataType, codec, resultCopy)
-                            exceptions.append(e)
-                        }
-                      } else
-                        exceptions.append(exceptions.head)
-                    case _ =>
-                      processTypesArray(dataType, codec.duplicate(), result)
-                      processTypesArray(dataType, codec, resultCopy)
+                  ???
+                }
+              case (x, _, C_END) if isArray && from.toInt > key.toInt =>
+                if (statementsList.size == 1) {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT) && !statementsList.head._1.isInstanceOf[KeyWithArrExpr]) {
+                    ???
+                  } else {
+                    ???
                   }
+                } else {
+                  ???
                 }
-              }
-            case (x, _, l) if isArray && (limitInf.toInt > x.toInt || l.toInt < x.toInt) =>
-              if (list.head._2.contains(C_DOUBLEDOT))
-                dataType match {
-                  case (D_BSONOBJECT | D_BSONARRAY) =>
-                    val size: Int = codec.getIntLE(codec.readerIndex())
-                    val buf1: ByteBuf = codec.readBytes(size)
-                    val buf2: ByteBuf = execStatementPatternMatch(buf1, list, f)
-                    result.writeBytes(buf2.duplicate())
-                    resultCopy.writeBytes(buf2.duplicate())
-                    buf1.release()
-                    buf2.release()
-                  case _ =>
-                    processTypesArray(dataType, codec.duplicate(), result)
-                    processTypesArray(dataType, codec, resultCopy)
+              case (x, _, l) if isArray && (from.toInt <= x.toInt && l.toInt >= x.toInt) =>
+                if (statementsList.size == 1) {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT) && !statementsList.head._1.isInstanceOf[KeyWithArrExpr]) {
+                    ???
+                  } else {
+                    ???
+                  }
+                } else {
+                  ???
                 }
-              else {
-                processTypesArray(dataType, codec.duplicate(), result)
-                processTypesArray(dataType, codec, resultCopy)
-              }
-            case (x, _, l) if !isArray =>
-              if (list.head._2.contains(C_DOUBLEDOT)) {
-                dataType match {
-                  case (D_BSONOBJECT | D_BSONARRAY) =>
-                    val size: Int = codec.getIntLE(codec.readerIndex())
-                    val buf1: ByteBuf = codec.readBytes(size)
-                    val buf2: ByteBuf = execStatementPatternMatch(buf1, list, f)
-                    result.writeBytes(buf2.duplicate())
-                    resultCopy.writeBytes(buf2.duplicate())
-                    buf1.release()
-                    buf2.release()
-                  case _ =>
-                    processTypesArray(dataType, codec.duplicate(), result)
-                    processTypesArray(dataType, codec, resultCopy)
+              case (x, _, l) if isArray && (from.toInt > x.toInt || l.toInt < x.toInt) =>
+                if (statementsList.size == 1) {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT) && !statementsList.head._1.isInstanceOf[KeyWithArrExpr]) {
+                    ???
+                  } else {
+                    ???
+                  }
+                } else {
+                  ???
                 }
-              } else throw CustomException("*ModifyArrayEnd* Not a Array")
-          }
+              case (x, _, l) if !isArray =>
+                if (statementsList.size == 1) {
+                  if (statementsList.head._2.contains(C_DOUBLEDOT) && !statementsList.head._1.isInstanceOf[KeyWithArrExpr]) {
+                    ???
+                  } else {
+                    ???
+                  }
+                } else {
+                  ???
+                }
+            }
+            ???
+        }
       }
     }
-    result.capacity(result.writerIndex())
-    resultCopy.capacity(resultCopy.writerIndex())
-    val a: ByteBuf = Unpooled.buffer(result.capacity() + 4).writeIntLE(result.capacity() + 4).writeBytes(result)
-    val b: ByteBuf = Unpooled.buffer(resultCopy.capacity() + 4).writeIntLE(resultCopy.capacity() + 4).writeBytes(resultCopy)
-    result.release()
-    resultCopy.release()
-    if (condition.equals(TO_RANGE))
-      if (exceptions.isEmpty) {
-        val bson: BosonImpl = new BosonImpl(byteArray = Option(a.array()))
-        a.release()
-        b.release()
-        bson
-      } else
-        throw exceptions.head
-    else if (condition.equals(UNTIL_RANGE)) {
 
-      if (exceptions.length <= 1) {
-        val bson: BosonImpl = new BosonImpl(byteArray = Option(b.array()))
-        a.release()
-        b.release()
-        bson
-      } else
-        throw exceptions.head
-    } else {
-
-      if (exceptions.isEmpty) {
-        val bson: BosonImpl = new BosonImpl(byteArray = Option(a.array()))
-        a.release()
-        b.release()
-        bson
-      } else
-        throw exceptions.head
+    val emptyCodec = codec.getCodecData match {
+      case Left(_) => CodecObject.toCodec(Unpooled.EMPTY_BUFFER)
+      case Right(_) => CodecObject.toCodec("")
     }
-  }
-
-  /**
-    * Function used to copy the values of an array that aren't of interest while searching for the last element
-    *
-    * @param list            A list with pairs that contains the key of interest and the type of operation
-    * @param fieldID         Name of the field of interest
-    * @param dataType        Type of the value found and processing
-    * @param codec           Structure from which we are reading the old values
-    * @param f               Function given by the user with the new value
-    * @param condition       Represents a type of injection, it can be END, ALL, FIRST, # TO #, # UNTIL #
-    * @param limitInf        Represent the inferior limit when a range is given
-    * @param limitSup        Represent the superior limit when a range is given
-    * @param resultCodec     Structure to where we write the values
-    * @param resultCopyCodec Auxiliary structure to where we write the values in case the previous cycle was the last one
-    * @tparam T Type of the value being injected
-    */
-  @deprecated
-  private def processTypesArrayEnd[T](list: List[(Statement, String)], fieldID: String, dataType: Int, codec: Codec, f: T => T, condition: String, limitInf: String = C_ZERO, limitSup: String = C_END, resultCodec: Codec, resultCopyCodec: Codec): (Codec, Codec) = {
-    dataType match {
-
-      case D_FLOAT_DOUBLE =>
-        val tokenRead = codec.readToken(SonNumber(CS_DOUBLE))
-        (resultCodec.writeToken(tokenRead), resultCopyCodec.writeToken(tokenRead))
-
-      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
-        val lengthToken = codec.readToken(SonNumber(CS_INTEGER))
-        val tokenRead = codec.readToken(SonString(CS_STRING))
-        val resultWithLength = resultCodec.writeToken(lengthToken)
-        val resultCopyWithLength = resultCopyCodec.writeToken(lengthToken)
-        (resultWithLength.writeToken(tokenRead), resultCopyWithLength.writeToken(tokenRead))
-
-      case D_BSONOBJECT =>
-        modifyArrayEndWithKey(list, codec, fieldID, f, condition, limitInf, limitSup, resultCodec, resultCopyCodec)
-
-      case D_BSONARRAY =>
-        modifyArrayEndWithKey(list, codec, fieldID, f, condition, limitInf, limitSup, resultCodec, resultCopyCodec)
-
-      case D_NULL => (codec, codec)
-
-      case D_INT =>
-        val tokenRead = codec.readToken(SonNumber(CS_INTEGER))
-        (resultCodec.writeToken(tokenRead), resultCopyCodec.writeToken(tokenRead))
-
-      case D_LONG =>
-        val tokenRead = codec.readToken(SonNumber(CS_LONG))
-        (resultCodec.writeToken(tokenRead), resultCopyCodec.writeToken(tokenRead))
-
-      case D_BOOLEAN =>
-        val tokenRead = codec.readToken(SonBoolean(CS_BOOLEAN))
-        (resultCodec.writeToken(tokenRead), resultCopyCodec.writeToken(tokenRead))
+    val codecWithoutSize = iterateDataStructure(emptyCodec)
+    val finalSize = codecWithoutSize.getCodecData match {
+      case Left(byteBuf) => byteBuf.capacity
+      case Right(string) => string.length
     }
+    emptyCodec.writeToken(emptyCodec, SonNumber(CS_INTEGER, finalSize)) + codecWithoutSize
   }
 
   /**
     * Function used to search for the last element of an array that corresponds to field with name fieldID
     *
-    * @param list            A list with pairs that contains the key of interest and the type of operation
-    * @param codec           Structure from which we are reading the old values
-    * @param fieldID         Name of the field of interest
-    * @param f               Function given by the user with the new value
-    * @param condition       Represents a type of injection, it can me END, ALL, FIRST, # TO #, # UNTIL #
-    * @param limitInf        Represent the inferior limit when a range is given
-    * @param limitSup        Represent the superior limit when a range is given
-    * @param resultCodec     Structure to where we write the values
-    * @param resultCopyCodec Auxiliary structure to where we write the values in case the previous cycle was the last one
-    * @tparam T Type of the value being injected
-    * @return a new BosonImpl with the new value injected
+    * @param statementsList
+    * @param codec
+    * @param fieldID
+    * @param injFunction
+    * @param condition
+    * @param from
+    * @param to
+    * @tparam T
+    * @return
     */
-  @deprecated
-  def modifyArrayEndWithKey[T](list: List[(Statement, String)], codec: Codec, fieldID: String, f: T => T, condition: String, limitInf: String = C_ZERO, limitSup: String = C_END, resultCodec: Codec, resultCopyCodec: Codec): (Codec, Codec) = {
-    /*
-    * Se fieldID for vazia, então deve ser chamada a funcao modifyArrayEnd to work on Root
-    *ByteBuf tem de ser duplicado no input
-    * */
-    val startReaderIndex: Int = codec.getReaderIndex
-    val originalSize: Int = codec.readSize
-    while ((codec.getReaderIndex - startReaderIndex) < originalSize) {
-      val dataType: Int = codec.readDataType
-      val modifiedResultCodec = resultCodec.writeToken(SonNumber(CS_BYTE, dataType))
-      val modifiedResultCopiedCodec = resultCopyCodec.writeToken(SonNumber(CS_BYTE, dataType))
-      dataType match {
-        case 0 =>
-        case _ =>
-          val (key, b): (String, Byte) = {
+  private def modifyArrayEndWithKey[T](statementsList: StatementsList, codec: Codec, fieldID: String, injFunction: T => T, condition: String, from: String, to: String = C_END): Codec = {
+    val startReaderIndex = codec.getReaderIndex
+    val originalSize = codec.readSize
 
-            val key: String = codec.readToken(SonString(CS_NAME)) match {
-              case SonString(_, keyStr) => keyStr.asInstanceOf[String]
+    val emptyDataStructure: Codec = codec.getCodecData match {
+      case Left(_) => CodecObject.toCodec(Unpooled.EMPTY_BUFFER)
+      case Right(_) => CodecObject.toCodec("")
+    }
+
+    def iterateDataStructure(currentCodec: Codec, currentCodecCopy: Codec): Codec = {
+      if ((codec.getReaderIndex - startReaderIndex) < originalSize) currentCodec
+      else {
+        val dataType: Int = codec.readDataType
+        val codecWithDataType: Codec = codec.writeToken(currentCodec, SonNumber(CS_BYTE, dataType))
+        val codecWithDataTypeCopy: Codec = codec.writeToken(currentCodecCopy, SonNumber(CS_BYTE, dataType))
+        dataType match {
+          case 0 => iterateDataStructure(codecWithDataType, codecWithDataTypeCopy)
+          case _ =>
+            val (key, byte): (String, Byte) = {
+              val key: String = codec.readToken(SonString(CS_NAME)) match {
+                case SonString(_, keyString) => keyString.asInstanceOf[String]
+              }
+              val token = codec.readToken(SonBoolean(C_ZERO))
+              val byte: Byte = token match {
+                case SonBoolean(_, byteBooelan) => byteBooelan.asInstanceOf[Byte]
+              }
+              (key, byte)
             }
-            val token = codec.readToken(SonBoolean(C_ZERO))
-            val b: Byte = token match {
-              case SonBoolean(_, res) => res.asInstanceOf[Byte]
-            }
-            (key, b)
-          }
-          val modResultCodec = modifiedResultCodec.writeToken(SonString(CS_STRING, key)).writeToken(SonNumber(CS_BYTE, b))
-          val modResultCopyCodec = modifiedResultCopiedCodec.writeToken(SonString(CS_STRING, key)).writeToken(SonNumber(CS_BYTE, b))
-          key match {
-            case extracted if (fieldID.toCharArray.deep == extracted.toCharArray.deep || isHalfword(fieldID, extracted)) && dataType == D_BSONARRAY =>
-              /*
-              * Found a field equal to key
-              * Perform Injection
-              * */
-              if (list.size == 1) {
-                if (list.head._2.contains(C_DOUBLEDOT)) {
-                  val partialCodec = codec.readToken(SonArray(CS_ARRAY)) match {
-                    case SonString(_, result) => CodecObject.toCodec(result)
-                  }
-                  modifyArrayEnd(list, partialCodec, f, condition, limitInf, limitSup)
+            val modResultCodec = codec.writeToken(codecWithDataType, SonString(CS_STRING, key))
+            val modResultCodecCopy = codec.writeToken(codecWithDataTypeCopy, SonString(CS_STRING, key))
 
-                } else modifyArrayEnd(list, codec, f, condition, limitInf, limitSup)
+            val resCodec = codec.writeToken(modResultCodec, SonString(CS_STRING, byte))
+            val resCodecCopy = codec.writeToken(modResultCodecCopy, SonString(CS_STRING, byte))
+            key match {
+              //In case we the extracted elem name is the same as the one we're looking for (or they're halfwords) and the
+              //datatype is a BsonArray
+              case extracted if (fieldID.toCharArray.deep == extracted.toCharArray.deep || isHalfword(fieldID, extracted)) && dataType == D_BSONARRAY =>
+                if (statementsList.size == 1) {
 
-              } else {
-                if (list.head._2.contains(C_DOUBLEDOT)) {
-                  val partialCodec = codec.readToken(SonArray(CS_ARRAY)) match {
-                    case SonString(_, result) => CodecObject.toCodec(result)
+                  if (statementsList.head._2.contains(C_DOUBLEDOT)) {
+                    val partialCodec = codec.readToken(SonArray(CS_ARRAY)) match {
+                      case SonArray(_, result) => result match {
+                        case byteBuf: ByteBuf => CodecObject.toCodec(byteBuf)
+                        case jsonString: String => CodecObject.toCodec(jsonString)
+                      }
+                    }
+                    val newCodec = modifyArrayEnd(statementsList, partialCodec, injFunction, condition, from, to)
+                    iterateDataStructure(newCodec, newCodec.duplicate)
+                  } else {
+                    val newCodec: Codec = modifyArrayEnd(statementsList, codec, injFunction, condition, from, to)
+                    iterateDataStructure(newCodec, newCodec.duplicate)
                   }
-                  //                  val size: Int = buffer.getIntLE(buffer.readerIndex())
-                  //                  val buf1: ByteBuf = buffer.readRetainedSlice(size)
-                  //                  // val buf2: ByteBuf = modifyArrayEnd(list, buf1, f, condition, limitInf, limitSup).getByteBuf
-                  //                  buf1.release()
-                  //                  //                  if (condition.equals(TO_RANGE))
-                  //                  //                  //  result.writeBytes(buf2)
-                  //                  //                  else if (condition.equals(UNTIL_RANGE))
-                  //                  //                 //   resultCopy.writeBytes(buf2)
-                  //                  //                  else {
-                  //                  //                 //   result.writeBytes(buf2)
-                  //                  //                  }
-                  //                  /// buf2.release()
 
                 } else {
-                  val res = modifyArrayEnd(list, codec, f, condition, limitInf, limitSup)
-                  //                  if (condition.equals(TO_RANGE))
-                  //                //    result.writeBytes(res.getByteBuf)
-                  //                  else if (condition.equals(UNTIL_RANGE))
-                  //                 //   resultCopy.writeBytes(res.getByteBuf)
-                  //                  else {
-                  //                 //   result.writeBytes(res.getByteBuf)
-                  //                  }
-                  //                //  res.getByteBuf.release()
-                }
-              }
-            case extracted if (fieldID.toCharArray.deep == extracted.toCharArray.deep || isHalfword(fieldID, extracted)) && dataType != D_BSONARRAY =>
-              if (list.head._2.contains(C_DOUBLEDOT) && list.head._1.isInstanceOf[KeyWithArrExpr])
-                processTypesArrayEnd(list, fieldID, dataType, codec, f, condition, limitInf, limitSup, resultCodec, resultCopyCodec)
-              else {
-                processTypesArray(dataType, codec.duplicate(), result) //TODO stopped here
-                processTypesArray(dataType, buffer, resultCopy)
-              }
-            case x if fieldID.toCharArray.deep != x.toCharArray.deep && !isHalfword(fieldID, x) =>
-              /*
-              * Didn't found a field equal to key
-              * Consume value and check deeper Levels
-              * */
-              if (list.head._2.contains(C_DOUBLEDOT) && list.head._1.isInstanceOf[KeyWithArrExpr]) {
-                processTypesArrayEnd(list, fieldID, dataType, buffer, f, condition, limitInf, limitSup, result, resultCopy)
-              } else {
-                processTypesArray(dataType, buffer.duplicate(), result)
-                processTypesArray(dataType, buffer, resultCopy)
-              }
-          }
-      }
-    }
-    result.capacity(result.writerIndex())
-    resultCopy.capacity(resultCopy.writerIndex())
-    val a: ByteBuf = Unpooled.buffer(result.capacity() + 4).writeIntLE(result.capacity() + 4).writeBytes(result)
-    val b: ByteBuf = Unpooled.buffer(resultCopy.capacity() + 4).writeIntLE(resultCopy.capacity() + 4).writeBytes(resultCopy)
-    result.release()
-    resultCopy.release()
-    if (condition.equals(TO_RANGE))
-      new BosonImpl(byteArray = Option(a.array()))
-    else if (condition.equals(UNTIL_RANGE))
-      new BosonImpl(byteArray = Option(b.array()))
-    else {
-      // //////println("condition END")
-      new BosonImpl(byteArray = Option(a.array()))
-    }
-  }
 
-  /**
-    * Function used to search for a element within another element
-    *
-    * @param list  A list with pairs that contains the key of interest and the type of operation
-    * @param codec Structure from which we are reading the old values
-    * @param key   Name of the field of interest
-    * @param elem  Name of elements to search inside the obejcts inside an Array
-    * @param f     Function given by the user with the new value
-    * @tparam T Type of the value being injected
-    * @return A structure with the new element injected
-    */
-  @deprecated
-  def modifyHasElem[T](list: List[(Statement, String)], codec: Codec, asd: String, elem: String, f: Function[T, T]): Codec = {
-    val startReader: Int = codec.getReaderIndex
-    val size: Int = codec.readSize
-    while ((codec.getReaderIndex - startReader) < size) {
-      val dataType: Int = codec.readDataType
-
-      val codecWithDataType = codec.writeToken(SonNumber(CS_BYTE, dataType))
-      dataType match {
-        case 0 =>
-        case _ =>
-          val (key, b): (String, Byte) = {
-            val key: String = codecWithDataType.readToken(SonString(CS_NAME)) match {
-              case SonString(_, keyString) => keyString.asInstanceOf[String]
-            }
-            val token = codecWithDataType.readToken(SonBoolean(C_ZERO))
-            val b: Byte = token match {
-              case SonBoolean(_, result) => result.asInstanceOf[Byte]
-            }
-            (key, b)
-          }
-          val codecWithKey = codecWithDataType.writeToken(SonString(CS_STRING, key)).writeToken(SonNumber(CS_BYTE, b))
-          key match {
-            case x if (key.toCharArray.deep == x.toCharArray.deep || isHalfword(key, x)) && dataType == D_BSONARRAY =>
-            // val newBuf: ByteBuf = searchAndModify(list, buf, elem, f).getByteBuf
-            // result.writeBytes(newBuf)
-            // newBuf.release()
-            case x if (key.toCharArray.deep == x.toCharArray.deep || isHalfword(key, x)) && (dataType != D_BSONARRAY) =>
-              if (list.head._2.contains(C_DOUBLEDOT))
-                processTypesHasElem(list, dataType, key, elem, buf, f, result)
-              else
-                processTypesArray(dataType, buf, result)
-            case x if key.toCharArray.deep != x.toCharArray.deep && !isHalfword(key, x) =>
-              if (list.head._2.contains(C_DOUBLEDOT))
-                processTypesHasElem(list, dataType, key, elem, buf, f, result)
-              else
-                processTypesArray(dataType, buf, result)
-          }
-      }
-    }
-    result.capacity(result.writerIndex())
-    val finalResult: ByteBuf = Unpooled.buffer(result.capacity() + 4).writeIntLE(result.capacity() + 4).writeBytes(result)
-    result.release()
-    finalResult
-  }
-
-  /**
-    * Function used to copy values that aren't of interest while searching for a element inside a object inside a array
-    *
-    * @param list     A list with pairs that contains the key of interest and the type of operation
-    * @param dataType Type of the value found and processing
-    * @param key      Name of the field of interest
-    * @param elem     Name of elements to search inside the objects inside an Array
-    * @param buf      Structure from which we are reading the old values
-    * @param f        Function given by the user with the new value
-    * @param result   Structure to where we write the values
-    * @tparam T Type of the value being injected
-    */
-  @deprecated
-  private def processTypesHasElem[T](list: List[(Statement, String)], dataType: Int, key: String, elem: String, buf: ByteBuf, f: (T) => T, result: ByteBuf): Unit = {
-    dataType match {
-      case D_FLOAT_DOUBLE =>
-        val value0: Double = buf.readDoubleLE()
-        result.writeDoubleLE(value0)
-      case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
-        val valueLength: Int = buf.readIntLE()
-        val bytes: ByteBuf = buf.readBytes(valueLength)
-        result.writeIntLE(valueLength)
-        result.writeBytes(bytes)
-        bytes.release()
-      case D_BSONOBJECT =>
-        val length: Int = buf.getIntLE(buf.readerIndex())
-        val bsonBuf: ByteBuf = buf.readBytes(length)
-        val newBsonBuf: ByteBuf = modifyHasElem(list, bsonBuf, key, elem, f)
-        bsonBuf.release()
-        result.writeBytes(newBsonBuf)
-        newBsonBuf.release()
-      case D_BSONARRAY =>
-        val length: Int = buf.getIntLE(buf.readerIndex())
-        val bsonBuf: ByteBuf = buf.readRetainedSlice(length)
-        val newBsonBuf: ByteBuf = modifyHasElem(list, bsonBuf, key, elem, f)
-        bsonBuf.release()
-        result.writeBytes(newBsonBuf)
-        newBsonBuf.release()
-      case D_NULL =>
-      case D_INT =>
-        val value0: Int = buf.readIntLE()
-        result.writeIntLE(value0)
-      case D_LONG =>
-        val value0: Long = buf.readLongLE()
-        result.writeLongLE(value0)
-      case D_BOOLEAN =>
-        val value0: Boolean = buf.readBoolean()
-        result.writeBoolean(value0)
-    }
-  }
-
-  /**
-    * Function used to search for an element inside a object inside a array after finding the key of interest
-    *
-    * @param list   A list with pairs that contains the key of interest and the type of operation
-    * @param buf    Structure from which we are reading the old values
-    * @param elem   Name of elements to search inside the objects inside an Array
-    * @param f      Function given by the user with the new value
-    * @param result Structure to where we write the values
-    * @tparam T Type of the value being injected
-    * @return a new BosonImpl with the new value injected
-    */
-  @deprecated
-  private def searchAndModify[T](list: List[(Statement, String)], buf: ByteBuf, elem: String, f: Function[T, T], result: ByteBuf = Unpooled.buffer()): BosonImpl = {
-    val startReader: Int = buf.readerIndex()
-    val size: Int = buf.readIntLE()
-    while ((buf.readerIndex() - startReader) < size) {
-      val dataType: Int = buf.readByte().toInt
-      result.writeByte(dataType)
-      dataType match {
-        case 0 =>
-        case _ =>
-          val (isArray, k, b): (Boolean, Array[Byte], Byte) = {
-            val key: ListBuffer[Byte] = new ListBuffer[Byte]
-            while (buf.getByte(buf.readerIndex()) != 0 || key.length < 1) {
-              val b: Byte = buf.readByte()
-              key.append(b)
-            }
-            val b: Byte = buf.readByte()
-            (key.forall(byte => byte.toChar.isDigit), key.toArray, b)
-          }
-          result.writeBytes(k).writeByte(b)
-          val keyString: String = new String(k)
-          dataType match {
-            case D_BSONOBJECT =>
-              val bsonSize: Int = buf.getIntLE(buf.readerIndex())
-              val bsonBuf: ByteBuf = Unpooled.buffer(bsonSize)
-              buf.getBytes(buf.readerIndex(), bsonBuf, bsonSize) //we're copying only a part we're interested in - in this case its the size of the object and the object it self
-            val hasElem: Boolean = hasElement(bsonBuf.duplicate(), elem)
-              if (hasElem) {
-                if (list.size == 1) {
-                  if (list.head._2.contains(C_DOUBLEDOT)) {
-                    val buf1: ByteBuf = buf.readBytes(bsonSize)
-                    val buf2: ByteBuf = Unpooled.buffer(buf1.capacity()).writeBytes(buf1)
-                    buf1.release()
-                    val bsonBytes: Array[Byte] = buf2.array()
-                    buf2.release()
-                    val newArray: Array[Byte] = applyFunction(f, bsonBytes).asInstanceOf[Array[Byte]]
-                    val newbuf: ByteBuf = Unpooled.buffer(newArray.length).writeBytes(newArray)
-                    val buf3: ByteBuf = execStatementPatternMatch(newbuf, list, f)
-                    newbuf.release()
-                    result.writeBytes(buf3)
-                    buf3.release()
-                    bsonBuf.release()
+                  if (statementsList.head._2.contains(C_DOUBLEDOT)) {
+                    val partialCodec = codec.readToken(SonArray(CS_ARRAY)) match {
+                      case SonArray(_, result) => result match {
+                        case byteBuf: ByteBuf => CodecObject.toCodec(byteBuf)
+                        case jsonString: String => CodecObject.toCodec(jsonString)
+                      }
+                    }
+                    ???
                   } else {
-                    val buf1: ByteBuf = buf.readBytes(bsonSize)
-                    val buf2: ByteBuf = Unpooled.buffer(buf1.capacity()).writeBytes(buf1)
-                    val array: Array[Byte] = buf2.array()
-                    buf2.release()
-                    val bsonBytes: Array[Byte] = array
-                    buf1.release()
-                    val value1: Array[Byte] = applyFunction(f, bsonBytes).asInstanceOf[Array[Byte]]
-                    result.writeBytes(value1)
-                    bsonBuf.release()
+                    val res = modifyArrayEnd(statementsList, codec, injFunction, condition, from, to)
+                    ???
                   }
+
+                }
+
+              //If we found the desired elem but the dataType is not an array, or if we didn't find the desired elem
+              case _ =>
+                if (statementsList.head._2.contains(C_DOUBLEDOT) && statementsList.head._1.isInstanceOf[KeyWithArrExpr]) {
+                  val (processedCodec, processedCodecCopy): (Codec, Codec) = ??? //processTypesArrayEnd(list, fieldID, dataType, codec, f, condition, limitInf, limitSup, resultCodec, resultCopyCodec)
+                  iterateDataStructure(processedCodec, processedCodecCopy)
                 } else {
-                  if (list.head._2.contains(C_DOUBLEDOT)) {
-                    val size: Int = buf.getIntLE(buf.readerIndex())
-                    val buf1: ByteBuf = buf.readBytes(size)
-                    val buf2: ByteBuf = execStatementPatternMatch(buf1, list.drop(1), f)
-                    val buf3: ByteBuf = execStatementPatternMatch(buf2.duplicate(), list, f)
-                    result.writeBytes(buf3)
-                    buf1.release()
-                    buf2.release()
-                    buf3.release()
-                    bsonBuf.release()
-                  } else {
-                    val buf1: ByteBuf = buf.readBytes(bsonSize)
-                    result.writeBytes(execStatementPatternMatch(bsonBuf, list.drop(1), f))
-                    buf1.release()
-                    bsonBuf.release()
-                  }
+                  //                  processTypesArray(dataType, codec.duplicate(), result)
+                  //                  processTypesArray(dataType, buffer, resultCopy)
+                  ???
                 }
-              } else {
-                val buf1: ByteBuf = buf.readBytes(bsonSize)
-                result.writeBytes(buf1)
-                buf1.release()
-              }
-            case _ =>
-              processTypesArray(dataType, buf, result)
-          }
-      }
-    }
-    result.capacity(result.writerIndex())
-    val finalResult: ByteBuf = Unpooled.buffer(result.capacity() + 4).writeIntLE(result.capacity() + 4).writeBytes(result)
-    result.release()
-    new BosonImpl(byteArray = Option(finalResult.array()))
-  }
 
-  /**
-    * Function used to see if an object contains a certain elem inside
-    *
-    * @param buf  Structure from which we are reading the old values
-    * @param elem Name of element to search inside the objects inside an Array
-    * @return A boolean meaning if the element has been found or not
-    */
-  @deprecated
-  def hasElement(buf: ByteBuf, elem: String): Boolean = {
-    val size: Int = buf.readIntLE()
-    val key: ListBuffer[Byte] = new ListBuffer[Byte]
-    while (buf.readerIndex() < size && (!elem.equals(new String(key.toArray)) && !isHalfword(elem, new String(key.toArray)))) {
-      key.clear()
-      val dataType: Int = buf.readByte()
-      dataType match {
-        case 0 =>
-        case _ =>
-          while (buf.getByte(buf.readerIndex()) != 0 || key.length < 1) {
-            val b: Byte = buf.readByte()
-            key.append(b)
-          }
-          buf.readByte()
-          dataType match {
-            case D_ZERO_BYTE =>
-            case D_FLOAT_DOUBLE =>
-              buf.readDoubleLE()
-            case D_ARRAYB_INST_STR_ENUM_CHRSEQ =>
-              val valueLength: Int = buf.readIntLE()
-              buf.readBytes(valueLength).release()
-            case D_BSONOBJECT =>
-              val valueLength: Int = buf.getIntLE(buf.readerIndex())
-              buf.readBytes(valueLength).release()
-            case D_BSONARRAY =>
-              val valueLength: Int = buf.getIntLE(buf.readerIndex())
-              buf.readBytes(valueLength).release()
-            case D_BOOLEAN =>
-              buf.readByte()
-            case D_NULL =>
-            case D_INT =>
-              buf.readIntLE()
-            case D_LONG =>
-              buf.readLongLE()
-            //case _ =>
-          }
-      }
-    }
-    new String(key.toArray).toCharArray.deep == elem.toCharArray.deep || isHalfword(elem, new String(key.toArray))
-  }
-
-  /**
-    * Function used to perform the search process recursively, performing the search key by key, separated by '.' or '..' and performing the injection when the final key is reached
-    *
-    * @param netty1     Structure from which we are reading the old values, can be either a ByteBuf or a string
-    * @param statements A list with pairs that contains the key of interest and the type of operation
-    * @param f          Function given by the user with the new value
-    * @tparam T Type of the value being injected
-    * @return A new structure with the final value injected
-    */
-  @deprecated
-  def execStatementPatternMatch[T](netty1: Either[ByteBuf, String], statements: List[(Statement, String)], f: Function[T, T]): Codec = {
-    val codec: Codec = netty1 match {
-      case Right(x) => CodecObject.toCodec(x)
-      case Left(x) => CodecObject.toCodec(x)
-    }
-
-    val statement: Statement = statements.head._1
-    val newStatementList: List[(Statement, String)] = statements
-
-    statement match {
-      case KeyWithArrExpr(key: String, arrEx: ArrExpr) =>
-        val input: (String, Int, String, Any) =
-          (arrEx.leftArg, arrEx.midArg, arrEx.rightArg) match {
-
-            case (i, o1, o2) if o1.isDefined && o2.isDefined =>
-              (key, arrEx.leftArg, o1.get.value, o2.get)
-
-            case (i, o1, o2) if o1.isEmpty && o2.isEmpty =>
-              (key, arrEx.leftArg, TO_RANGE, arrEx.leftArg)
-
-            case (0, str, None) =>
-              str.get.value match {
-                case C_FIRST =>
-                  (key, 0, TO_RANGE, 0)
-                case C_END =>
-                  (key, 0, C_END, None)
-                case C_ALL =>
-                  (key, 0, TO_RANGE, C_END)
-              }
-          }
-        val res: ByteBuf = execArrayFunction(newStatementList, codec, f, input._1, input._2, input._3, input._4)
-        val finalResult: ByteBuf = resultByteBuf.writeBytes(res) //começar a alterar aqui
-        res.release()
-        finalResult.capacity(finalResult.writerIndex())
-
-
-      case ArrExpr(leftArg: Int, midArg: Option[RangeCondition], rightArg: Option[Any]) =>
-        val input: (String, Int, String, Any) =
-          (leftArg, midArg, rightArg) match {
-            case (i, o1, o2) if midArg.isDefined && rightArg.isDefined =>
-              (EMPTY_KEY, leftArg, midArg.get.value, rightArg.get)
-            case (i, o1, o2) if midArg.isEmpty && rightArg.isEmpty =>
-              (EMPTY_KEY, leftArg, TO_RANGE, leftArg)
-            case (0, str, None) =>
-              str.get.value match {
-                case C_FIRST => (EMPTY_KEY, 0, TO_RANGE, 0)
-                case C_END => (EMPTY_KEY, 0, C_END, None)
-                case C_ALL => (EMPTY_KEY, 0, TO_RANGE, C_END)
-              }
-          }
-        val res: ByteBuf = execArrayFunction(newStatementList, netty1, f, input._1, input._2, input._3, input._4)
-        val finalResult: ByteBuf = resultByteBuf.writeBytes(res).capacity(resultByteBuf.writerIndex())
-        res.release()
-        finalResult
-
-      case HalfName(half: String) => modifyAll(newStatementList, codec, half, f)
-
-      case HasElem(key: String, elem: String) => modifyHasElem(newStatementList, codec, key, elem, f)
-
-      case Key(key: String) => modifyAll(newStatementList, codec, key, f)
-
-      case ROOT =>
-        val res: ByteBuf = execRootInjection(codec, f)
-        resultByteBuf.writeBytes(res)
-        res.release()
-        resultByteBuf.capacity(resultByteBuf.writerIndex())
-
-      case _ => throw CustomException("Wrong Statements, Bad Expression.")
-      // Never Gets Here
-
-    }
-  }
-
-  /**
-    * Function used with 'execStatementPatternMatch' to perform the process of recursive search while dealing with arrays
-    *
-    * @param list   A list with pairs that contains the key of interest and the type of operation
-    * @param buf    Structure from which we are reading the old values
-    * @param f      Function given by the user with the new value
-    * @param key    Name of value to be used in search (can be empty)
-    * @param left   Left argument of the array conditions
-    * @param mid    Middle argument of the array conditions
-    * @param right  Right argument of array conditions
-    * @param result Structure to where we write the values
-    * @tparam T Type of the value being injected
-    * @return A new structure with the final value injected
-    **/
-  @deprecated
-  def execArrayFunction[T](list: List[(Statement, String)], codec: Codec, f: Function[T, T], key: String, left: Int, mid: String, right: Any, result: ByteBuf = Unpooled.buffer()): ByteBuf = {
-    (key, left, mid.toLowerCase(), right) match {
-      case (EMPTY_KEY, 0, C_END, None) =>
-        val size: Int = buf.getIntLE(buf.readerIndex())
-        val buf1: ByteBuf = buf.readBytes(size)
-        val res: ByteBuf = ??? //modifyArrayEnd(list, buf1, f, C_END, 0.toString).getByteBuf
-        result.writeBytes(res)
-        res.release()
-        buf1.release()
-        result.capacity(result.writerIndex())
-      case (EMPTY_KEY, a, UNTIL_RANGE, C_END) =>
-        val size: Int = buf.getIntLE(buf.readerIndex())
-        val buf1: ByteBuf = buf.readBytes(size)
-        val res: ByteBuf = ??? //modifyArrayEnd(list, buf1, f, UNTIL_RANGE, a.toString).getByteBuf
-        result.writeBytes(res)
-        res.release()
-        buf1.release()
-        result.capacity(result.writerIndex())
-      case (EMPTY_KEY, a, TO_RANGE, C_END) => // "[# .. end]"
-        val size: Int = buf.getIntLE(buf.readerIndex())
-        val buf1: ByteBuf = buf.readRetainedSlice(size)
-        val res: ByteBuf = ??? //modifyArrayEnd(list, buf1, f, TO_RANGE, a.toString).getByteBuf
-        buf1.release()
-        result.writeBytes(res)
-        res.release()
-        result.capacity(result.writerIndex())
-      case (EMPTY_KEY, a, expr, b) if b.isInstanceOf[Int] =>
-        expr match {
-          case TO_RANGE =>
-            val size: Int = buf.getIntLE(buf.readerIndex())
-            val buf1: ByteBuf = buf.readRetainedSlice(size)
-            val res: ByteBuf = ??? // modifyArrayEnd(list, buf1, f, TO_RANGE, a.toString, b.toString).getByteBuf
-            buf1.release()
-            result.writeBytes(res)
-            res.release()
-            result.capacity(result.writerIndex())
-          case UNTIL_RANGE =>
-            val size: Int = buf.getIntLE(buf.readerIndex())
-            val buf1: ByteBuf = buf.readRetainedSlice(size)
-            val res: ByteBuf = ??? //modifyArrayEnd(list, buf1, f, UNTIL_RANGE, a.toString, b.toString).getByteBuf
-            buf1.release()
-            result.writeBytes(res)
-            res.release()
-            result.capacity(result.writerIndex())
+            }
         }
-      case (k, 0, C_END, None) =>
-        val size: Int = buf.getIntLE(buf.readerIndex())
-        val buf1: ByteBuf = buf.readBytes(size)
-        val res: ByteBuf = ??? // modifyArrayEndWithKey(list, buf1, k, f, C_END, 0.toString).getByteBuf
-        result.writeBytes(res)
-        res.release()
-        buf1.release()
-        result.capacity(result.writerIndex())
-
-      case (k, a, UNTIL_RANGE, C_END) =>
-        val size: Int = buf.getIntLE(buf.readerIndex())
-        val buf1: ByteBuf = buf.readRetainedSlice(size)
-        val res: ByteBuf = ??? //modifyArrayEndWithKey(list, buf1, k, f, UNTIL_RANGE, a.toString).getByteBuf
-        buf1.release()
-        result.writeBytes(res)
-        res.release()
-        result.capacity(result.writerIndex())
-      case (k, a, TO_RANGE, C_END) =>
-        val size: Int = buf.getIntLE(buf.readerIndex())
-        val buf1: ByteBuf = buf.readRetainedSlice(size)
-        val res: ByteBuf = ??? //modifyArrayEndWithKey(list, buf1, k, f, TO_RANGE, a.toString).getByteBuf
-        buf1.release()
-        result.writeBytes(res)
-        res.release()
-        result.capacity(result.writerIndex())
-      case (k, a, expr, b) if b.isInstanceOf[Int] =>
-        expr match {
-          case TO_RANGE =>
-            val size: Int = buf.getIntLE(buf.readerIndex())
-            val buf1: ByteBuf = buf.readRetainedSlice(size)
-            val res: ByteBuf = ??? //modifyArrayEndWithKey(list, buf1, k, f, TO_RANGE, a.toString, b.toString).getByteBuf
-            buf1.release()
-            result.writeBytes(res)
-            res.release()
-            result.capacity(result.writerIndex())
-          case UNTIL_RANGE =>
-            val size: Int = buf.getIntLE(buf.readerIndex())
-            val buf1: ByteBuf = buf.readRetainedSlice(size)
-            val res: ByteBuf = ??? //modifyArrayEndWithKey(list, buf1, k, f, UNTIL_RANGE, a.toString, b.toString).getByteBuf
-            buf1.release()
-            result.writeBytes(res)
-            res.release()
-            result.capacity(result.writerIndex())
-        }
+      }
+      iterateDataStructure(emptyDataStructure, emptyDataStructure.duplicate)
     }
+
+    //TODO - Check Either to create empty codec
+    val emptyCodec = CodecObject.toCodec(Unpooled.EMPTY_BUFFER)
+    val codecWithoutSize = iterateDataStructure(emptyCodec, emptyCodec)
+    val finalSize = codecWithoutSize.getCodecData match {
+      case Left(byteBuf) => byteBuf.capacity
+      case Right(string) => string.length
+    }
+    emptyCodec.writeToken(emptyCodec, SonNumber(CS_INTEGER, finalSize)) + codecWithoutSize
   }
 
   /**
-    * Function used when the injection is done at Root level
+    * Helper function to retrieve a codec with the key information written in it , and the key that was written
     *
-    * @param buffer Structure from which we are reading the old values
-    * @param f      Function given by the user with the new value
-    * @tparam T Type of the value being injected
-    * @return A new structure with the final value injected
+    * @param codec
+    * @return
     */
-  @deprecated
-  def execRootInjection[T](buffer: ByteBuf, f: Function[T, T]): ByteBuf = {
-    val bsonBytes: Array[Byte] = buffer.array()
-    val newBson: Array[Byte] = applyFunction(f, bsonBytes).asInstanceOf[Array[Byte]]
-    Unpooled.buffer(newBson.length).writeBytes(newBson)
+  private def writeKeyAndByte(codec: Codec, writableCodec: Codec): (Codec, String) = {
+    val key: String = codec.readToken(SonString(CS_NAME)) match {
+      case SonString(_, keyString) => keyString.asInstanceOf[String]
+    }
+    val b: Byte = codec.readToken(SonBoolean(C_ZERO)) match { //TODO FOR CodecJSON we cant read a boolean, we need to read an empty string
+      case SonBoolean(_, result) => result.asInstanceOf[Byte]
+    }
+
+    val codecWithKey = codec.writeToken(writableCodec, SonString(CS_STRING, key))
+    (codec.writeToken(codecWithKey, SonNumber(CS_BYTE, b)), key)
   }
 }
