+++ boson-core/src/main/scala/io/zink/boson/bson/codec/impl/CodecJson.scala
@@ -623,5 +623,12 @@ class CodecJson(str: String) extends Codec {
     val duplicate = str
     Right(duplicate)
   }
+
+  /**
+    *
+    * @param sumCodec
+    * @return
+    */
+  override def +(sumCodec: Codec): Codec = ???
 }
