+++ boson-scala/src/main/scala/io/zink/boson/impl/BosonExtractor.scala
@@ -3,7 +3,7 @@ package io.zink.boson.impl
 import java.nio.ByteBuffer
 
 import io.zink.boson.Boson
-import io.zink.boson.bson.bsonImpl.{BosonImpl, BosonImpl2}
+import io.zink.boson.bson.bsonImpl.BosonImpl
 import io.zink.boson.bson.bsonPath._
 import shapeless.TypeCase
 
@@ -15,15 +15,15 @@ import scala.concurrent.Future
   * These types doesn't compile with an implicit LabelledGeneric and for that reason a different class
   * is required.
   *
-  * @param expression String given by the User designated as BsonPath.
-  * @param extractFunction  Extract function given by the User to be applied after extraction.
+  * @param expression      String given by the User designated as BsonPath.
+  * @param extractFunction Extract function given by the User to be applied after extraction.
   * @tparam T Type of Value to be extracted.
   */
 class BosonExtractor[T](expression: String, extractFunction: T => Unit)(implicit tp: Option[TypeCase[T]]) extends Boson {
 
   private val boson: BosonImpl = new BosonImpl()
 
-  private val interpreter: Interpreter[T] = new Interpreter[T](boson,expression, fExt = Option(extractFunction))
+  private val interpreter: Interpreter[T] = new Interpreter[T](boson, expression, fExt = Option(extractFunction))
 
   //private val interpreter2 = new Interpreter[T](boson, expression, fExt = Option(extractFunction))
 
@@ -32,17 +32,17 @@ class BosonExtractor[T](expression: String, extractFunction: T => Unit)(implicit
     * statements used to instantiate the Interpreter.
     */
   // byteArr as argument to go to interpreter, Either[byte[],String]
-  private def runInterpreter(bsonEncoded: Either[Array[Byte],String]): Unit = {
+  private def runInterpreter(bsonEncoded: Either[Array[Byte], String]): Unit = {
     interpreter.run(bsonEncoded)
   }
 
-//  private def runInterpreter2(bsonEncoded: Seq[ByteBuf]): Any = {
-//    val results =
-//    bsonEncoded.par.map{ elem =>
-//      interpreter2.runExtractors(Left(elem), interpreter2.keyList,interpreter2.limitList)
-//    }.seq.flatten.map{ case e: ByteBuf => e.array()}
-//    extractFunction(results.asInstanceOf[T])
-//  }
+  //  private def runInterpreter2(bsonEncoded: Seq[ByteBuf]): Any = {
+  //    val results =
+  //    bsonEncoded.par.map{ elem =>
+  //      interpreter2.runExtractors(Left(elem), interpreter2.keyList,interpreter2.limitList)
+  //    }.seq.flatten.map{ case e: ByteBuf => e.array()}
+  //    extractFunction(results.asInstanceOf[T])
+  //  }
 
   override def go(bsonByteEncoding: Array[Byte]): Future[Array[Byte]] = {
     val future: Future[Array[Byte]] = Future {
@@ -52,22 +52,22 @@ class BosonExtractor[T](expression: String, extractFunction: T => Unit)(implicit
     future
   }
 
-//
-//  def go2(encodedStructures: Seq[ByteBuf]): Future[Seq[ByteBuf]] = {
-//    val future: Future[Seq[ByteBuf]] =
-//      Future {
-//        runInterpreter2(encodedStructures)
-//        encodedStructures
-//
-//      }
-//    future
-//  }
+  //
+  //  def go2(encodedStructures: Seq[ByteBuf]): Future[Seq[ByteBuf]] = {
+  //    val future: Future[Seq[ByteBuf]] =
+  //      Future {
+  //        runInterpreter2(encodedStructures)
+  //        encodedStructures
+  //
+  //      }
+  //    future
+  //  }
 
   override def go(bsonByteBufferEncoding: ByteBuffer): Future[ByteBuffer] = {
     val future: Future[ByteBuffer] =
       Future {
         runInterpreter(Left(bsonByteBufferEncoding.array()))
-          bsonByteBufferEncoding
+        bsonByteBufferEncoding
       }
     future
   }
